% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agregion.lines}
\alias{agregion.lines}
\title{Grainbelt inland border}
\format{
...
}
\source{
copied from the SSF: unknown -- originally sourced or created by Fiona Evans.
}
\usage{
agregion.lines
}
\description{
Set of line segments that together make up the inland border of the south-west
WA grainbelt region.
** format needs sorting out.
}
\keyword{datasets}
