% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agregion_img}
\alias{agregion_img}
\title{Agricultural areas image mask}
\format{
a 381 * 321 matrix consisting of zeroes/ones.
}
\source{
copied from the SSF: unknown -- originally sourced or created by
Fiona Evans.
}
\usage{
agregion_img
}
\description{
Image mask to restrict mapping to agricultural areas of south west Western
Australia, including both the grainbelt and the lower south-west corner.
}
\keyword{datasets}
