% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coast.img}
\alias{coast.img}
\title{Ocean image mask}
\format{
a 381 * 321 matrix consisting of zeroes/ones.
}
\source{
copied from the SSF: unknown -- originally sourced or created by Fiona Evans
}
\usage{
coast.img
}
\description{
Image mask to restrict mapping to land areas of south west Western Australia
}
\keyword{datasets}
