% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coast.lines}
\alias{coast.lines}
\title{South west WA coast line}
\format{
...
}
\source{
copied from the SSF: unknown -- originally sourced or created by Fiona Evans.
}
\usage{
coast.lines
}
\description{
Set of line segments that together make up the WA coast line.
** currently includes some boundary lines on the ocean -- plan is to get rid of those
** format needs to be sorted, and saved as something else.
}
\keyword{datasets}
