% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{color_df}
\alias{color_df}
\title{Apply colours to vector}
\usage{
color_df(x, col_df, trunc = TRUE)
}
\arguments{
\item{x}{Input vector}

\item{col_df}{Data frame of discrete colours to use}

\item{trunc}{Should the range be truncated to within the range of the colour
data frame? If true (default), values less than the minimum will be mapped
as the minimum while values greater than the maximum will be mapped as the
maximum. If false, values below minimum or above maximum will be set to NA.
This is set to TRUE, to allow for correct implementation of legacy calls to
the function.}
}
\description{
Returns a vector of colours corresponding to the numerical values of input
vector. Colour sets can be one of the inbuilt sets (eg.
\code{mapcol_deciles}) or can be created using \code{make_col_set}
}
\keyword{internal}
