% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{convert_krig_to_color}
\alias{convert_krig_to_color}
\title{convert the kriged values to colours}
\usage{
convert_krig_to_color(im, color)
}
\arguments{
\item{im}{object created by the 'map_krig' internal function}

\item{color}{mapping colour set; should be a data.frame. Can be created using
\code{make_col_set}}
}
\description{
convert the kriged values to colours
}
\keyword{internal}
