% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_climograph.R
\name{create_climograph}
\alias{create_climograph}
\title{Creates a climograph of weather data from the DPIRD API}
\usage{
create_climograph(weather)
}
\arguments{
\item{weather}{A list of weather station data from the \acronym{DPIRD}
Science \acronym{API}.}
}
\value{
A \CRANpkg{ggplot2} object
}
\description{
Creates a climograph of weather data from the DPIRD API
}
\examples{
create_climograph(weather = weather)

}
\author{
Adam H. Sparks, \email{adam.sparks@dpird.wa.gov.au}
}
