% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_seq_check.R
\name{date_seq_check}
\alias{date_seq_check}
\title{Check dates for sequentialness and completeness}
\usage{
date_seq_check(x, y, jan1 = TRUE)
}
\arguments{
\item{x}{A vector of dates}

\item{y}{A vector of values with location name}
}
\value{
A NULL value
}
\description{
Invisibly returns a \code{NULL} value if no errors are found, else stops with
an informative message providing the user with missing date values in the
sequence.
}
\examples{

x <- seq(as.Date("2021-01-01"), as.Date("2021-01-31"), by = "1 day")
y <- "BA"
date_seq_check(x = x, y = y)
}
\author{
Adam H. Sparks, adam.sparks@dpird.wa.gov.au
}
