% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_silo_points.R
\name{.query_silo}
\alias{.query_silo}
\title{Construct and send SILO API queries}
\usage{
.query_silo(station_id, latitude, longitude)
}
\arguments{
\item{longitude}{}
}
\description{
Construct and send SILO API queries
}
