% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dpird.logo}
\alias{dpird.logo}
\title{DPIRD logo}
\format{
raster
}
\source{
version in use in the SSF copied; original source forgotten
}
\usage{
dpird.logo
}
\description{
Black and white logo for the Department of Primary Industries and Regional Development.
Current as of May 2021.
}
\keyword{datasets}
