% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpird_extreme_weather_values.R
\docType{data}
\name{dpird_extreme_weather_values}
\alias{dpird_extreme_weather_values}
\title{Extreme weather data values available from the DPIRD Weather 2.0 'API'}
\format{
A \link{vector} object of 57 items.
}
\source{
\url{https://www.agric.wa.gov.au/weather-api-20}
}
\usage{
dpird_extreme_weather_values
}
\description{
A \link{vector} object containing 57 items representing valid values to supply
to \code{get_extreme_weather()}'s \code{values} argument.
}
\keyword{datasets}
