% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpird_minute_values.R
\docType{data}
\name{dpird_minute_values}
\alias{dpird_minute_values}
\title{Minute weather data values available from the DPIRD Weather 2.0 'API'}
\format{
A \link{vector} object of 12 items.
}
\source{
\url{https://www.agric.wa.gov.au/weather-api-20}
}
\usage{
dpird_minute_values
}
\description{
A \link{vector} object containing 12 items representing valid values to supply
to \code{get_dpird_minute()}'s \code{values} argument.
}
\keyword{datasets}
