% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dpird_office_towns}
\alias{dpird_office_towns}
\title{Localities with a DPIRD office}
\format{
...
}
\source{
create based on current office locations
\url{https://www.agric.wa.gov.au/office-locations}.
}
\usage{
dpird_office_towns
}
\description{
Default set of office locations to be added region maps
}
\keyword{datasets}
