% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpird_summary_values.R
\docType{data}
\name{dpird_summary_values}
\alias{dpird_summary_values}
\title{Summary weather data values available from the DPIRD Weather 2.0 'API'}
\format{
A \link{vector} object of 75 items.
}
\source{
\url{https://www.agric.wa.gov.au/weather-api-20}
}
\usage{
dpird_summary_values
}
\description{
A \link{vector} object containing 75 items representing valid values to supply
to \code{get_dpird_summary()}'s \code{values} argument.
}
\keyword{datasets}
