% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{embed_legend}
\alias{embed_legend}
\title{Add legend}
\usage{
embed_legend(top, left, col, u = 0.5, cex = NULL, percent = FALSE)
}
\arguments{
\item{top}{Position of vertical centre of the top box of the legend. Top of
the legend will be at top + u/2}

\item{left}{Position of left edge of legend.}

\item{col}{Data frame of discrete colours to use. If this does not contain a
column called 'label', then the minimum and maximum values will be used to
determine the breaks, and the values will be added at the breaks. ie. there
will be one more value than there are colours.}

\item{u}{Scaling parameter for legend boxes (Default = 0.5). Boxes will be
square with sides 'u' long.}

\item{cex}{Scaling parameter for the text. If this is not provided, it will
be calculated as u/2, for backwards compatibility.}

\item{percent}{Should the legend values be converted to percentages? If true,
the values are multiplied by 100 and '\%' added. (Default = FALSE)}
}
\description{
Embed a legend in a plot. (vertical legend with top-left at left, top + u/2)
Note that this version is a combination of two pre-existing versions. One
version had a single scaling parameter, the second had two.
**There may still be some optimising to be had here.
}
