% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep_for_forecast_towns.R
\name{find_forecast_towns}
\alias{find_forecast_towns}
\title{Find nearest BOM forecast towns for a given location}
\usage{
find_forecast_towns(latitude = -35.3, longitude = 149.2, distance_km = 100)
}
\arguments{
\item{latitude}{A \code{numeric} value of latitude in decimal degree (DD)
format. By default, Canberra (approximately).}

\item{longitude}{A \code{numeric} value of longitude in decimal degree (DD)
format. By default, Canberra (approximately).}

\item{distance_km}{A \code{numeric} value of the distance in kilometres from the
\code{latitude} and \code{longitude} point beyond which values will not be returned.}
}
\value{
A \code{\link{data.table}} of all forecast towns (in this package)
sorted by distance from \var{latlon}, ascending.
}
\description{
Find nearest BOM forecast towns for a given location
}
\examples{

# sweep for forecast towns near Esperance, WA
find_forecast_towns(lat = -33.8614, lon = 121.8913)

}
\seealso{
Other bomrang-ported: 
\code{\link{get_available_radar}()},
\code{\link{get_precis_forecast}()},
\code{\link{parse_precis_forecast}()},
\code{\link{update_forecast_towns}()},
\code{\link{update_station_locations}()}
}
\author{
Hugh Parsonage, \email{hugh.parsonage@gmail.com} and
James Goldie, \email{me@jamesgoldie.dev} and Adam Sparks,
\email{adam.sparks@dpird.wa.gov.au}
}
\concept{bomrang-ported}
