% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nearby_silo_stations.R
\name{find_nearby_silo_stations}
\alias{find_nearby_silo_stations}
\title{Find stations within a given radius of a geographic point or weather station}
\usage{
find_nearby_silo_stations(
  distance_km = NULL,
  longitude = NULL,
  latitude = NULL,
  station_id = NULL
)
}
\arguments{
\item{distance_km}{A \code{numeric} value for the distance in kilometres in which
to search for nearby stations from a given geographic location or known
\code{station_id}. Cannot be used when \code{longitude} and \code{latitude} are provided.}

\item{longitude}{A \code{numeric} value for longitude expressed as decimal degrees
(DD) (WGS84). Required if \code{latitude} is used. Cannot be used if \code{station_id}
is provided.}

\item{latitude}{A \code{numeric} value for latitude expressed as decimal degrees
(DD) (WGS84). Required if \code{longitude} is used. Cannot be used if
\code{station_id} is provided.}

\item{station_id}{A \code{character} value for a known station from which the
nearby stations are to be identified.}
}
\value{
A \code{data.table} of available stations and associated metadata.
}
\description{
Searches for stations available in the SILO database and returnes values
sorted from nearest to farthest.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# find stations within 50 km of Finke Post Office, NT
find_nearby_silo_stations(distance_km = 50,
                         station_id = "015526")

# find stations within 500 km of Esperance, WA
find_nearby_silo_stations(distance_km = 50,
                          longitude = 121.8913,
                          latitude = -33.8614)
\dontshow{\}) # examplesIf}
}
