% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nearby_stations.R
\name{find_nearby_stations}
\alias{find_nearby_stations}
\title{Find Nearby Stations}
\usage{
find_nearby_stations(
  latitude = NULL,
  longitude = NULL,
  station_code = NULL,
  distance_km,
  api_key = NULL,
  which_api = "silo"
)
}
\arguments{
\item{latitude}{A \code{numeric} value for latitude expressed as decimal degrees
(DD) (WGS84).  Optional and defaults to \code{NULL}.  Required if \code{longitude}
is not \code{NULL.}}

\item{longitude}{A \code{numeric} value for longitude expressed as decimal degrees
(DD) (WGS84).  Optional and defaults to \code{NULL}.  Required if \code{latitude} is
not \code{NULL}.}

\item{station_code}{A \code{string} with the station code for the station of
interest.  Optional and defaults to \code{NULL}.}

\item{distance_km}{A \code{numeric} value for distance to limit the search from
the station or location of interest.  Defaults to 100 km.}

\item{api_key}{A \code{string} value that is the user's \acronym{API} key from
\acronym{DPIRD} (see \url{https://www.agric.wa.gov.au/web-apis}).  Only used
when \code{which_api} is 'DPIRD' or 'all'.}

\item{which_api}{A \code{string} value that indicates which \acronym{API} to use.
Defaults to "silo" only. Valid values are "all", for both \acronym{SILO}
(\acronym{BOM}) and \acronym{DPIRD} weather station networks; "silo" for
only stations in the \acronym{SILO} network; or "dpird" for stations in the
\acronym{DPIRD} network.}
}
\value{
a \link[data.table:data.table]{data.table::data.table} with "station_code", "station_name",
"latitude", "longitude", "elev_m", "state", "owner", and "distance".
Data are sorted by increasing distance from station or location of
interest.
}
\description{
Find nearby weather stations given geographic coordinates or a station code.
Either a combination of \code{latitude} and \code{longitude} or \code{station_code} must
be provided.
}
\note{
You can request your own \acronym{API} key from \acronym{DPIRD} for
free by filling out the form found at
\url{https://www.agric.wa.gov.au/web-apis}.
}
\examples{
\dontrun{
# Query WA only stations and return both DPIRD's and BOM's stations for
# the Northam WA station, returning stations with 50 km of this station
wa_stn <- find_nearby_stations(
  station_code = "NO",
  distance_km = 50,
  api_key = "your_api_key",
  which_api = "DPIRD"
)

# Query Wagga Wagga BOM station.
wagga_stn <- find_nearby_stations(
  latitude = -35.1583,
  longitude = 147.4575,
  distance_km = 200,
  which_api = "SILO"
)
}

}
\seealso{
Other DPIRD: 
\code{\link{dpird_extreme_weather_values}},
\code{\link{dpird_minute_values}},
\code{\link{dpird_summary_values}},
\code{\link{get_dpird_availability}()},
\code{\link{get_dpird_extremes}()},
\code{\link{get_dpird_minute}()},
\code{\link{get_dpird_summaries}()},
\code{\link{get_station_metadata}()}

Other SILO: 
\code{\link{get_data_drill}()},
\code{\link{get_patched_point}()},
\code{\link{get_station_metadata}()},
\code{\link{silo_daily_values}}
}
\author{
Rodrigo Pires, \email{rodrigo.pires@dpird.wa.gov.au} and Adam H. Sparks
\email{adam.sparks@dpird.wa.gov.au}
}
\concept{DPIRD}
\concept{SILO}
