% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ag_bulletin.R
\name{get_ag_bulletin}
\alias{get_ag_bulletin}
\title{Get BOM agriculture bulletin information for select stations}
\usage{
get_ag_bulletin(state = "AUS")
}
\arguments{
\item{state}{Australian state or territory as full name or postal code.
Fuzzy string matching via \code{\link[base]{agrep}} is done.  Defaults to
\dQuote{AUS} returning all state bulletins, see Details for more.}
}
\value{
A data frame as a \code{\link[data.table]{data.table}} object of Australia
\acronym{BOM} agricultural bulletin information.
}
\description{
Fetch the \acronym{BOM} agricultural bulletin information and return it in a
data frame.  Ported from \pkg{bomrang}.
}
\details{
Allowed state and territory postal codes, only one state per request
or all using \code{AUS}.
\describe{
\item{ACT}{Australian Capital Territory (will return NSW)}
\item{NSW}{New South Wales}
\item{NT}{Northern Territory}
\item{QLD}{Queensland}
\item{SA}{South Australia}
\item{TAS}{Tasmania}
\item{VIC}{Victoria}
\item{WA}{Western Australia}
\item{AUS}{Australia, returns forecast for all states, NT and ACT}
}
}
\note{
Data and Information Use
Please note the copyright notice and disclaimer,
\url{http://www.bom.gov.au/other/copyright.shtml} related to the use of this
information. Users of this information are deemed to have read and
accepted the conditions described therein.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

ag_bulletin <- get_ag_bulletin(state = "QLD")
ag_bulletin
\dontshow{\}) # examplesIf}
}
\references{
Agricultural observations are retrieved from the Australian Bureau of
Meteorology (\acronym{BOM}) Weather Data Services Agriculture Bulletins, \cr
\url{http://www.bom.gov.au/catalogue/observations/about-agricultural.shtml}

and

Australian Bureau of Meteorology (\acronym{BOM})) Weather Data Services
Observation of Rainfall, \cr
\url{http://www.bom.gov.au/climate/how/observations/rain-measure.shtml}

Station location and other metadata are sourced from the Australian Bureau of
Meteorology (\acronym{BOM}) webpage, Bureau of Meteorology Site Numbers: \cr
\url{http://www.bom.gov.au/climate/cdo/about/site-num.shtml}
}
\seealso{
\link{parse_ag_bulletin}

Other bomrang-ported: 
\code{\link{find_forecast_towns}()},
\code{\link{get_available_imagery}()},
\code{\link{get_available_radar}()},
\code{\link{get_coastal_forecast}()},
\code{\link{get_precis_forecast}()},
\code{\link{get_satellite_imagery}()},
\code{\link{parse_coastal_forecast}()},
\code{\link{parse_precis_forecast}()},
\code{\link{update_ag_station_locations}()},
\code{\link{update_forecast_towns}()}
}
\author{
Adam H. Sparks, \email{adamhsparks@dpird.wa.gov.au} and Paul Melloy
\email{paul@melloy.com.au}
}
\concept{bomrang-ported}
