% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coastal_forecast.R
\name{get_coastal_forecast}
\alias{get_coastal_forecast}
\title{Get BOM coastal waters forecast}
\usage{
get_coastal_forecast(state = "AUS")
}
\arguments{
\item{state}{Australian state or territory as full name or postal code.
Fuzzy string matching via \code{\link[base:agrep]{base::agrep()}} is done.  Defaults to "AUS"
returning all state forecasts, see details for further information.}
}
\value{
A \link[data.table:data.table]{data.table::data.table} of an Australia \acronym{BOM} Coastal Waters
Forecast.
}
\description{
Fetch the \acronym{BOM} daily Coastal Waters Forecast for a specified state
or region.
}
\details{
Allowed state and territory postal codes, only one state per request
or all using 'AUS':
\describe{
\item{ACT}{Australian Capital Territory (will return NSW)}
\item{NSW}{New South Wales}
\item{NT}{Northern Territory}
\item{QLD}{Queensland}
\item{SA}{South Australia}
\item{TAS}{Tasmania}
\item{VIC}{Victoria}
\item{WA}{Western Australia}
\item{AUS}{Australia, returns forecast for all states, NT and ACT}
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

coastal_forecast <- get_coastal_forecast(state = "NSW")
coastal_forecast
\dontshow{\}) # examplesIf}
}
\references{
Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data
Services \cr \url{http://www.bom.gov.au/catalogue/data-feeds.shtml}.

And also,

Location data and other metadata come from the \acronym{BOM} anonymous
\acronym{FTP} server with spatial data \cr
\url{ftp://ftp.bom.gov.au/anon/home/adfd/spatial/}, specifically the
\acronym{DBF} file portion of a shapefile, \cr
\url{ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf}.
}
\seealso{
\link{parse_coastal_forecast}

Other bomrang-ported: 
\code{\link{find_forecast_towns}()},
\code{\link{get_ag_bulletin}()},
\code{\link{get_available_imagery}()},
\code{\link{get_available_radar}()},
\code{\link{get_precis_forecast}()},
\code{\link{get_satellite_imagery}()},
\code{\link{parse_coastal_forecast}()},
\code{\link{parse_precis_forecast}()},
\code{\link{update_forecast_towns}()}
}
\author{
Dean Marchiori, \email{deanmarchiori@gmail.com} and Paul Melloy
\email{paul@melloy.com.au}
}
\concept{bomrang-ported}
