% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_drill_apsim.R
\name{get_data_drill_apsim}
\alias{get_data_drill_apsim}
\title{Get DataDrill Weather Data in the APSIM Format From SILO}
\usage{
get_data_drill_apsim(
  longitude,
  latitude,
  start_date,
  end_date = Sys.Date(),
  api_key
)
}
\arguments{
\item{longitude}{A single \code{numeric} value  representing the longitude of the
point-of-interest.}

\item{latitude}{A single \code{numeric} value representing the latitude of the
point-of-interest.}

\item{start_date}{A \code{character} string or \code{Date} object representing the
beginning of the range to query in the format \dQuote{yyyy-mm-dd}
(ISO8601).  Data returned is inclusive of this range.}

\item{end_date}{A \code{character} string or \code{Date} object representing the end of
the range query in the format  \dQuote{yyyy-mm-dd} (ISO8601).  Data
returned is inclusive of this range.  Defaults to the current system date.}

\item{api_key}{A \code{character }string specifying a valid email address to use
for the request.  The query will return an error if a valid email address
is not provided.}
}
\value{
An \link{apsimx} object of class \sQuote{met} with attributes.
}
\description{
Fetch \acronym{APSIM} .met file formatted weather data from the weather data
from the \acronym{SILO} \acronym{API} of spatially interpolated weather
data (DataDrill).  The daily climate surfaces have been derived either by
splining or kriging the observational data.  The returned values contain
\dQuote{source} columns, which denote how the observations were derived.
The grid spans 112° to 154°, -10° to -44° with resolution 0.05° latitude by
0.05° longitude (approximately 5 km × 5 km).
}
\section{Value information}{


Solar radiation: total incoming downward shortwave radiation on a horizontal
surface, derived from estimates of cloud oktas and sunshine duration^3.

Relative humidity: calculated using the vapour pressure measured at 9am, and
the saturation vapour pressure computed using either the maximum or minimum
temperature^6.

Evaporation and evapotranspiration: an overview of the variables provided by
\acronym{SILO} is available here,
\url{https://data.longpaddock.qld.gov.au/static/publications/Evapotranspiration_overview.pdf}.
}

\section{Data codes}{

Data codes
Where possible (depending on the file format), the data are supplied with
codes indicating how each datum was obtained.

\tabular{rl}{
\strong{Code}\tab \strong{Source}\cr
\strong{0}:\tab Official observation as supplied by the Bureau of
Meteorology.\cr
\strong{15}:\tab Deaccumulated rainfall (original observation was recorded
over a period exceeding the standard 24 hour observation period).\cr
\strong{25}:\tab Interpolated from daily observations for that date.\cr
\strong{26}:\tab Synthetic Class A pan evaporation, calculated from
temperatures, radiation and vapour pressure.\cr
\strong{35}:\tab Interpolated from daily observations using an anomaly
interpolation method.\cr
\strong{75}:\tab Interpolated from the long term averages of daily
observations for that day of year.\cr
}
}

\examples{
\dontrun{
# requires an API key as your email address
# Source data from latitude and longitude coordinates (gridded data) for
# max and minimum temperature and rainfall for Southwood, QLD.
wd <- get_data_drill_apsim(
  latitude = -27.85,
  longitude = 150.05,
  start_date = "20221001",
  end_date = "20221201",
  api_key = "your@email"
)
}

}
\references{
\enumerate{
\item Rayner, D. (2005). Australian synthetic daily Class A pan evaporation.
Technical Report December 2005, Queensland Department of Natural Resources
and Mines, Indooroopilly, Qld., Australia, 40 pp.
\item Morton, F. I. (1983). Operational estimates of areal evapotranspiration
and their significance to the science and practice of hydrology, \emph{Journal
of Hydrology}, Volume 66, 1-76.
\item Zajaczkowski, J., Wong, K., & Carter, J. (2013). Improved historical
solar radiation gridded data for Australia, \emph{Environmental Modelling &
Software}, Volume 49, 64–77. DOI: \doi{10.1016/j.envsoft.2013.06.013}.
\item Food and Agriculture Organization of the United Nations,
Irrigation and drainage paper 56: Crop evapotranspiration - Guidelines for
computing crop water requirements, 1998.
\item ASCE’s Standardized Reference Evapotranspiration Equation, proceedings of
the National Irrigation Symposium, Phoenix, Arizona, 2000.
\item For further details refer to Jeffrey, S.J., Carter, J.O., Moodie, K.B. and
Beswick, A.R. (2001). Using spatial interpolation to construct a
comprehensive archive of Australian climate data, \emph{Environmental Modelling
and Software}, Volume 16/4, 309-330. DOI:
\doi{10.1016/S1364-8152(01)00008-1}.
}
}
\seealso{
Other SILO: 
\code{\link{find_nearby_stations}()},
\code{\link{get_data_drill}()},
\code{\link{get_patched_point_apsim}()},
\code{\link{get_patched_point}()},
\code{\link{get_station_metadata}()},
\code{\link{silo_daily_values}}

Other data fetching: 
\code{\link{get_ag_bulletin}()},
\code{\link{get_coastal_forecast}()},
\code{\link{get_data_drill}()},
\code{\link{get_dpird_extremes}()},
\code{\link{get_dpird_minute}()},
\code{\link{get_dpird_summaries}()},
\code{\link{get_patched_point_apsim}()},
\code{\link{get_patched_point}()},
\code{\link{get_precis_forecast}()},
\code{\link{get_radar_imagery}()},
\code{\link{get_satellite_imagery}()}
}
\author{
Rodrigo Pires, \email{rodrigo.pires@dpird.wa.gov.au}, and Adam
Sparks, \email{adam.sparks@dpird.wa.gov.au}
}
\concept{SILO}
\concept{data fetching}
