% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multi_silo_points.R
\name{get_multi_silo_points_v2}
\alias{get_multi_silo_points_v2}
\title{Retrieve data from SILO (Scientific Information for Land Owners) API
for multiple sites}
\usage{
get_multi_silo_points_v2(
  station_id = NULL,
  latitude = NULL,
  longitude = NULL,
  first = NULL,
  last = NULL,
  data_format = "standard",
  email = NULL
)
}
\arguments{
\item{station_id}{An integer, station number available at the \acronym{SILO}
the network.}

\item{latitude}{A vector, representing the latitude of a point-of-interest}

\item{longitude}{A vector, representing the longitude of a point-of-interest}

\item{first}{A string representing the start date of the query in the
format 'yyyymmdd'}

\item{last}{A string representing the end date of the query in the
format 'yyyymmdd'}

\item{data_format}{A string specifying the type of data to retrieve.
Limited to 'alldata', 'monthly' or 'apsim'. Note 'apsim' and 'alldata'
retrieve daily data.}

\item{email}{A string specifying the email address to use for the request.
The query will return an error if a valid email address is not provided.}
}
\value{
A \code{list} object of \code{data.frames} with the retrieved data. Each
list item is a \code{data.frame} with data for a given station code or
geographic coordinates. New columns are added in both cases, \code{station_id}
or \code{latitude} and \code{longitude},  respectively.
}
\description{
This function is a wrapper around \code{get_silo_points()} to handle
query with multiple sites, either station codes or latitude and longitude
coordinates. See help for \code{get_silo_points()} for details.
}
\examples{
# start multisession
future::plan("multisession")
# Query multiple stations (BoM, by code)
my_stations <- c(65030, 89033, 39083, 8061)
wd <- get_multi_silo_points(station_id = my_stations,
                            first = "20220401",
                            last = "20221001",
                            data_format = "monthly",
                            email = "YOUR_EMAIL_ADDRESS")

# Query multiple stations (using latitude and longitude coordinates)
mylat <- c(-33.512, -30.665, -27.309, -24.237)
mylon <- c(115.821, 117.487, 119.649, 121.584)
wd <- get_multi_silo_points(latitude = mylat,
                            longitude = mylon,
                            first = "202101201",
                            last = "20220228",
                            data_format = "alldata",
                            email = "YOUR_EMAIL_ADDRESS")
}
\seealso{
Other SILO: 
\code{\link{get_silo_points}()}
}
\concept{SILO}
