% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nearby_stations.R
\name{get_nearby_stations}
\alias{get_nearby_stations}
\title{Get a list of nearby weather stations from coordinates or station ID code}
\usage{
get_nearby_stations(
  latitude = NULL,
  longitude = NULL,
  site = NULL,
  distance_km = 100,
  api_key = NULL,
  wa_only = TRUE,
  dpird_only = FALSE
)
}
\arguments{
\item{latitude}{Latitude expressed as decimal degrees (DD) (WGS84).}

\item{longitude}{Longitude expressed as decimal degrees (DD) (WGS84).}

\item{site}{A string with the station ID code for the station of interest.
Optional and defaults to NULL.}

\item{distance_km}{Distance to limit the search from station of interest.
Defaults to 100 km.}

\item{api_key}{User's \acronym{API} key from \acronym{DPIRD}
(https://www.agric.wa.gov.au/web-apis)}

\item{wa_only}{Return stations in Western Australia only. Defaults to TRUE.}

\item{dpird_only}{Return only DPIRD owned stations and defaults to FALSE.}
}
\value{
a \code{data.frame} with stationCode, stationName, latitude, longitude,
state, owner and distance. Data is sorted by increasing distance from station
of interest.
}
\description{
Get a list of nearby weather stations from coordinates or station ID code
}
\examples{
# You must have an DPIRD API key to proceed
my_key <- rstudioapi::askForSecret()

# Query WA only stations and return both DPIRDs and BOMs stations.
my_station <- "NO"    # Northam
my_distance <- 50    # in km

wa_stn <- get_nearby_stations(
  site = my_station,
  distance_km = my_distance,
  api_key = my_key,
  dpird_only = FALSE,
  wa_only = TRUE
)

# Query Wagga Wagga BOM station.
# Need to set `dpird_only` and `wa_only` to `FALSE`
lat <- -35.1583
long <- 147.4575

wagga_stn <- get_nearby_stations(
  latitude = lat,
  longitude = long,
  distance_km = my_distance,
  api_key = mykey,
  dpird_only = FALSE,
  wa_only = FALSE
)

}
\author{
Rodrigo Pires, rodrigo.pires@dpird.wa.gov.au
}
