% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_patched_point.R
\encoding{UTF-8}
\name{get_patched_point}
\alias{get_patched_point}
\title{Get PatchedPoint Weather Data From SILO}
\usage{
get_patched_point(
  station_code,
  start_date,
  end_date = Sys.Date(),
  values = "all",
  api_key
)
}
\arguments{
\item{station_code}{A \code{character} string of the \acronym{BOM} station code
for the station of interest.}

\item{start_date}{A \code{character} string or \code{Date} object representing the
beginning of the range to query in the format \dQuote{yyyy-mm-dd}
(ISO8601).  Data returned is inclusive of this date.}

\item{end_date}{A \code{character} string or \code{Date} object representing the end of
the range query in the format  \dQuote{yyyy-mm-dd} (ISO8601).  Data
returned is inclusive of this date.  Defaults to the current system date.}

\item{values}{A \code{character} string with the type of weather data to
return.  See \strong{Available Values} for a full list of valid values.
Defaults to \code{all} with all available values being returned.}

\item{api_key}{A \code{character} string specifying a valid email address to use
for the request.  The query will return an error if a valid email address
is not provided.}
}
\value{
a \link{data.table} with the weather data queried with the
weather variables in alphabetical order.  The first eight columns will
always be:
\itemize{
\item \code{station_code},
\item \code{station_name},
\item \code{longitude},
\item \code{latitude},
\item \code{elev_m} (elevation in metres),
\item \code{date} (ISO8601 format, "YYYYMMDD"),
\item \code{year},
\item \code{month},
\item \code{day},
\item \code{extracted} (the date on which the query was made)
}
}
\description{
Fetch nicely formatted weather data from the \acronym{SILO} \acronym{API}
derived from the \acronym{BOM} station observations (PatchedPoint) data.
}
\section{Column Name Details}{
Column names are converted from the default returns of the API to be
snake_case formatted and where appropriate, the names of the values that
are analogous between \acronym{SILO} and \acronym{DPIRD} data are named
using the same name for ease of interoperability, \emph{e.g.}, using
\code{rbind()} to create a \code{data.table} that contains data from both APIs.

The \acronym{SILO} documentation provides the following information
for the PatchedPoint data.

\emph{These data are a continuous daily time series of data at either recording
stations or grid points across Australia:}
\itemize{
\item \emph{Data at station locations consists of observational records which have
been supplemented by interpolated estimates when observed data are missing.
Datasets are available at approximately 8,000 Bureau of Meteorology
recording stations around Australia.}
\item \emph{Data at grid points consists entirely of interpolated estimates. The
data are taken from the SILO gridded datasets and are available at any
pixel on a 0.05° × 0.05° grid over the land area of Australia (including
some islands).}
}
}

\section{Available Values}{


\describe{
\item{all}{Which will return all of the following values}
\item{rain (mm)}{Rainfall}
\item{max_temp (degrees C)}{Maximum temperature}
\item{min_temp (degrees C)}{Minimum temperature}
\item{vp (hPa)}{Vapour pressure}
\item{vp_deficit (hPa)}{Vapour pressure deficit}
\item{evap_pan (mm)}{Class A pan evaporation}
\item{evap_syn (mm)}{Synthetic
\ifelse{html}{\out{estimate<sup>1</sup>}}{estimate\eqn{^1}}}
\item{evap_comb (mm)}{Combination (synthetic estimate pre-1970, class A pan
1970 onwards)}
\item{evap_morton_lake (mm)}{Morton's shallow lake evaporation}
\item{radiation (Mj/\ifelse{html}{\out{m<sup>2</sup>}}{m\eqn{^2}})}{Solar
exposure, consisting of both direct and diffuse components}
\item{rh_tmax (\%)}{Relative humidity at the time of maximum temperature}
\item{rh_tmin (\%)}{Relative humidity at the time of minimum temperature}
\item{et_short_crop (mm)}{
\ifelse{html}{\out{FAO56<sup>4</sup>}}{FAO56\eqn{^4}} short crop}
\item{et_tall_crop (mm)}{
\ifelse{html}{\out{ASCE<sup>5</sup>}}{ASCE\eqn{^5}} tall
\ifelse{html}{\out{crop<sup>6</sup>}}{crop\eqn{^6}}}
\item{et_morton_actual (mm)}{Morton's areal actual evapotranspiration}
\item{et_morton_potential (mm)}{Morton's point potential evapotranspiration}
\item{et_morton_wet (mm)}{Morton's wet-environment areal potential
evapotranspiration over land}
\item{mslp (hPa)}{Mean sea level pressure}
}
}

\section{Value information}{


Solar radiation: total incoming downward shortwave radiation on a horizontal
surface, derived from estimates of cloud oktas and sunshine
\ifelse{html}{\out{duration<sup>3</sup>}}{duration\eqn{^3}}.

Relative humidity: calculated using the vapour pressure measured at 9am, and
the saturation vapour pressure computed using either the maximum or minimum
\ifelse{html}{\out{temperature<sup>6</sup>}}{temperature\eqn{^6}}.

Evaporation and evapotranspiration: an overview of the variables provided by
\acronym{SILO} is available here,
\url{https://data.longpaddock.qld.gov.au/static/publications/Evapotranspiration_overview.pdf}.
}

\section{Data codes}{


The data are supplied with codes indicating how each datum was obtained.

\describe{
\item{0}{Official observation as supplied by the Bureau of Meteorology}
\item{15}{Deaccumulated rainfall (original observation was recorded
over a period exceeding the standard 24 hour observation period).}
\item{25}{Interpolated from daily observations for that date.}
\item{26}{Synthetic Class A pan evaporation, calculated from
temperatures, radiation and vapour pressure.}
\item{35}{Interpolated from daily observations using an anomaly
interpolation method.}
\item{75}{Interpolated from the long term averages of daily
observations for that day of year.}
}
}

\examples{
\dontrun{
# requires an API key as your email address
# Source observation data for station Wongan Hills station, WA (008137)
wd <- get_patched_point(station_code = "008137",
               start_date = "2021-06-01",
               end_date = "2021-07-01",
               values = "all",
               api_key = "your_api_key")
}
}
\references{
\enumerate{
\item Rayner, D. (2005). Australian synthetic daily Class A pan evaporation.
Technical Report December 2005, Queensland Department of Natural Resources
and Mines, Indooroopilly, Qld., Australia, 40 pp.
\item Morton, F. I. (1983). Operational estimates of areal evapotranspiration
and their significance to the science and practice of hydrology, \emph{Journal
of Hydrology}, Volume 66, 1-76.
\item Zajaczkowski, J., Wong, K., & Carter, J. (2013). Improved historical
solar radiation gridded data for Australia, \emph{Environmental Modelling &
Software}, Volume 49, 64–77. DOI: \doi{10.1016/j.envsoft.2013.06.013}.
\item Food and Agriculture Organization of the United Nations,
Irrigation and drainage paper 56: Crop evapotranspiration - Guidelines for
computing crop water requirements, 1998.
\item ASCE’s Standardized Reference Evapotranspiration Equation, proceedings of
the National Irrigation Symposium, Phoenix, Arizona, 2000.
\item For further details refer to Jeffrey, S.J., Carter, J.O., Moodie, K.B. and
Beswick, A.R. (2001). Using spatial interpolation to construct a
comprehensive archive of Australian climate data, \emph{Environmental Modelling
and Software}, Volume 16/4, 309-330. DOI:
\doi{10.1016/S1364-8152(01)00008-1}.
}
}
\seealso{
Other SILO: 
\code{\link{find_nearby_stations}()},
\code{\link{get_data_drill_apsim}()},
\code{\link{get_data_drill}()},
\code{\link{get_patched_point_apsim}()},
\code{\link{get_station_metadata}()},
\code{\link{silo_daily_values}}

Other data fetching: 
\code{\link{get_ag_bulletin}()},
\code{\link{get_coastal_forecast}()},
\code{\link{get_data_drill_apsim}()},
\code{\link{get_data_drill}()},
\code{\link{get_dpird_apsim}()},
\code{\link{get_dpird_extremes}()},
\code{\link{get_dpird_minute}()},
\code{\link{get_dpird_summaries}()},
\code{\link{get_patched_point_apsim}()},
\code{\link{get_precis_forecast}()},
\code{\link{get_radar_imagery}()},
\code{\link{get_satellite_imagery}()}
}
\author{
Rodrigo Pires, \email{rodrigo.pires@dpird.wa.gov.au}, and Adam
Sparks, \email{adam.sparks@dpird.wa.gov.au}
}
\concept{SILO}
\concept{data fetching}
