% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_patched_point_apsim.R
\name{get_patched_point_apsim}
\alias{get_patched_point_apsim}
\title{Get PatchedPoint Weather Data in the APSIM Format From SILO}
\usage{
get_patched_point_apsim(
  station_code,
  start_date,
  end_date = Sys.Date(),
  api_key
)
}
\arguments{
\item{station_code}{A \code{character} string of the \acronym{BOM} station code
for the station of interest.}

\item{start_date}{A \code{character} string or \code{Date} object representing the
beginning of the range to query in the format \dQuote{yyyy-mm-dd}
(ISO8601).  Data returned is inclusive of this range.}

\item{end_date}{A \code{character} string or \code{Date} object representing the end of
the range query in the format  \dQuote{yyyy-mm-dd} (ISO8601).  Data
returned is inclusive of this range.  Defaults to the current system date.}

\item{api_key}{A \code{character} string specifying a valid email address to use
for the request.  The query will return an error if a valid email address
is not provided.}
}
\value{
An \link{apsimx} object of class \sQuote{met} with attributes.
}
\description{
Fetch \acronym{APSIM} .met file formatted weather data from the
\acronym{SILO} \acronym{API} derived from the \acronym{BOM} station
observations (PatchedPoint) data.
}
\details{
The \acronym{SILO} documentation provides the following information
for the PatchedPoint data.

\emph{These data are a continuous daily time series of data at either recording
stations or grid points across Australia:}
\itemize{
\item \emph{Data at station locations consists of observational records which have
been supplemented by interpolated estimates when observed data are missing.
Datasets are available at approximately 8,000 Bureau of Meteorology
recording stations around Australia.}
\item \emph{Data at grid points consists entirely of interpolated estimates. The
data are taken from the SILO gridded datasets and are available at any
pixel on a 0.05° × 0.05° grid over the land area of Australia (including
some islands).}
}
}
\section{Value information}{


Solar radiation: total incoming downward shortwave radiation on a horizontal
surface, derived from estimates of cloud oktas and sunshine duration^3.

Relative humidity: calculated using the vapour pressure measured at 9am, and
the saturation vapour pressure computed using either the maximum or minimum
temperature^6.

Evaporation and evapotranspiration: an overview of the variables provided by
\acronym{SILO} is available here,
\url{https://data.longpaddock.qld.gov.au/static/publications/Evapotranspiration_overview.pdf}.
}

\examples{
\dontrun{
# requires an API key as your email address
# Source observation data for station Wongan Hills station, WA (008137)
wd <- get_patched_point_apsim(
  station_code = "008137",
  start_date = "2021-06-01",
  end_date = "2021-07-01",
  api_key = "your@email"
)
}
}
\references{
\enumerate{
\item Rayner, D. (2005). Australian synthetic daily Class A pan evaporation.
Technical Report December 2005, Queensland Department of Natural Resources
and Mines, Indooroopilly, Qld., Australia, 40 pp.
\item Morton, F. I. (1983). Operational estimates of areal evapotranspiration
and their significance to the science and practice of hydrology, \emph{Journal
of Hydrology}, Volume 66, 1-76.
\item Zajaczkowski, J., Wong, K., & Carter, J. (2013). Improved historical
solar radiation gridded data for Australia, \emph{Environmental Modelling &
Software}, Volume 49, 64–77. DOI: \doi{10.1016/j.envsoft.2013.06.013}.
\item Food and Agriculture Organization of the United Nations,
Irrigation and drainage paper 56: Crop evapotranspiration - Guidelines for
computing crop water requirements, 1998.
\item ASCE’s Standardized Reference Evapotranspiration Equation, proceedings of
the National Irrigation Symposium, Phoenix, Arizona, 2000.
\item For further details refer to Jeffrey, S.J., Carter, J.O., Moodie, K.B. and
Beswick, A.R. (2001). Using spatial interpolation to construct a
comprehensive archive of Australian climate data, \emph{Environmental Modelling
and Software}, Volume 16/4, 309-330. DOI:
\doi{10.1016/S1364-8152(01)00008-1}.
}
}
\seealso{
Other SILO: 
\code{\link{find_nearby_stations}()},
\code{\link{get_data_drill_apsim}()},
\code{\link{get_data_drill}()},
\code{\link{get_patched_point}()},
\code{\link{get_station_metadata}()},
\code{\link{silo_daily_values}}

Other data fetching: 
\code{\link{get_ag_bulletin}()},
\code{\link{get_coastal_forecast}()},
\code{\link{get_data_drill_apsim}()},
\code{\link{get_data_drill}()},
\code{\link{get_dpird_extremes}()},
\code{\link{get_dpird_minute}()},
\code{\link{get_dpird_summaries}()},
\code{\link{get_patched_point}()},
\code{\link{get_precis_forecast}()},
\code{\link{get_radar_imagery}()},
\code{\link{get_satellite_imagery}()}
}
\author{
Rodrigo Pires, \email{rodrigo.pires@dpird.wa.gov.au}, and Adam
Sparks, \email{adam.sparks@dpird.wa.gov.au}
}
\concept{SILO}
\concept{data fetching}
