% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_silo_points.R
\name{get_silo_points}
\alias{get_silo_points}
\title{Retrieve data from SILO (Scientifc Information for Land Owners) API}
\usage{
get_silo_points(
  station_id = NULL,
  latitude = NULL,
  longitude = NULL,
  start_date = NULL,
  end_date = NULL,
  data_format = "alldata",
  email = NULL
)
}
\arguments{
\item{station_id}{An integer, station number available at the SILO
the network.}

\item{latitude}{A vector, represeting the latitude of a point-of-interest}

\item{longitude}{A vector, represeting the longitude of a point-of-interest}

\item{start_date}{A string representing the start date of the query in the
format 'yyyymmdd'}

\item{end_date}{A string representing the end date of the query in the
format 'yyyymmdd'}

\item{data_format}{A string specifying the type of data to retrieve.
Limited to 'alldata', 'monthly' or 'apsim'. Note 'apsim' and 'alldata'
retrieve daily data.}

\item{email}{A string specifying the email address to use for the request.
The query will return an error if a valid email address is not provided.}
}
\value{
A dataframe containing the retrieved data from the SILO API.
}
\description{
Download weather data from the SILO API from both station
observations (DataDrill) and gridded data (PatchedPointData). There are
three formats available: 'alldata' and 'apsim' with daily frequency and
'monthly' with, that's right, monthly frequency.
}
\examples{
# Source observation data for station Wongan Hills station, WA (8137)
wd <- get_silo_points(station_id = 8137,
                      start_date = "20210601",
                      end_date = "20210701",
                      data_type = "alldata",
                      email = "your@email")

# Source data from latitude and longitude coordinates (gridded data)
# Southwood, WLD in the 'apsim' format.
wd <-
get_silo_points(latitude = -27.85,
                longitude = 150.05,
                start_date = "20221001",
                end_date = "20221201",
                data_format = "apsim",
                email = "your@email")
}
