% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{make_col_set}
\alias{make_col_set}
\title{Make colour sets}
\usage{
make_col_set(col, cuts, key = NULL, sep = "-")
}
\arguments{
\item{col}{set of colours in any format recognised by R. One colour per
grouping}

\item{cuts}{set of break points. Must contain one more item than the list of
colours. \code{-Inf} and \code{Inf} are valid options for the minimum and
maximum values, respectively.}

\item{key}{if a non-standard set of labels are required for the key, they can
be supplied. The default set just uses the break points as the ranges.}

\item{sep}{Separator between number in key; default is hyphen}
}
\description{
Create a colour set for use in \code{map_weather} from a set of n colours
and a set of n + 1 cut points. It is assumed that the lowest and highest
values in the break points are set so that the bottom and top segments are <
the second value and > second last value, respectively. There are multiple
'default' colour sets available.
}
