% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{map_krig}
\alias{map_krig}
\title{Internal kriging functions}
\usage{
map_krig(data, varname, lambda, theta = 1, lat_lims, long_lims)
}
\arguments{
\item{data}{data frame containing the variable to be krigged (identified by
'varname') and latitude and longitude variables, which provide the spatial
structure to the data.}

\item{varname}{String indicating which column is to be krigged.}

\item{lambda}{Smoothing parameter for the Kriging. Larger values will give
greater smoothing}

\item{theta}{Kriging parameter (from the fields::Krig help file, the 'range
parameter'). Default is 1.}

\item{lat_lims}{latitudinal limits of mapping}

\item{long_lims}{longitudinal limits of mapping}
}
\description{
A set of internal functions that are called by various mapping functions (as
of writing there is only the one, which does the whole SWWA).
}
\keyword{internal}
