% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{map_krig_layer}
\alias{map_krig_layer}
\title{wrapper for map.krig}
\usage{
map_krig_layer(
  data,
  varname,
  lambda,
  lat_lims,
  long_lims,
  col,
  agregion,
  coast
)
}
\arguments{
\item{data}{data frame containing the variable to be krigged (identified by
'varname') and latitude and longitude variables, which provide the spatial
structure to the data.}

\item{varname}{String indicating which column is to be krigged.}

\item{lambda}{Smoothing parameter for the Kriging. Larger values will give
greater smoothing}

\item{lat_lims}{latitudinal limits of mapping}

\item{long_lims}{longitudinal limits of mapping}

\item{col}{specified colour data set needed to convert the continuous
values to the categorical colour map}

\item{agregion}{required mask to remove mapping of the land outside the
ag region}

\item{coast}{required mask to remove mapping of the ocean}
}
\description{
this is the interim function between map.ssf and map.krig
so that the map layer is generated.
}
