% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{map_weather}
\alias{map_weather}
\title{Gradient map of South-West WA.}
\usage{
map_weather(
  data,
  varname,
  col_df,
  percent,
  title = "",
  subtitle1 = "",
  subtitle2 = "",
  subtitle3 = NULL,
  dots = NULL,
  dots_label = NULL,
  plot_type = 0,
  name = NULL,
  html_path = NULL,
  mask_agregion = NULL,
  mask_coast = NULL,
  lines_shire = NULL,
  lines_coast = NULL,
  lines_agregion = NULL,
  logo = NULL,
  scale = 10,
  lambda = 0.01,
  legendx = 122.25,
  transparency = FALSE,
  box = TRUE,
  type = "krig"
)
}
\arguments{
\item{data}{Data frame with columns for the latitude and longitude locations
of the points, and the measurement of \code{varname} at those points. Other
columns will be ignored.}

\item{varname}{Name of variable to map}

\item{col_df}{Data frame containing discrete colour information. For
\code{map_ssf_dots} the defaults is NULL. When this is left as NULL, then a
rainbow of colours is applied to the unique values. Note that this may give
strange results if the variable to be plotted is a factor and the factor
order is not in the desired legend order.}

\item{percent}{Should the colour scale be displayed in percentages? If this
is true, the legend values will be assumed to be percentages, and the
percent symbol appended, unless labels are included in the \code{col_df}}

\item{title}{Main title for the map. This will be the first line of text in
the finished map. Blank by default}

\item{subtitle1}{The first line of the map subtitle. Blank by default}

\item{subtitle2}{The second line of the map subtitle. Blank by default}

\item{subtitle3}{Third subtitle for the map, if required.}

\item{dots}{Data frame containing the latitude and longitude of any points to
plot. It is assumed that these will have the variable names
\code{latitude} and \code{longitude}. (not case sensitive?). If type is
'point' and this parameter is provided, then only the points that are in
both will be plotted.}

\item{dots_label}{Data frame containing the latitude, longitude, and text for
any labelled points to plot. Note that this does not have to be the same
set as the \code{dots}; inclusion of the latitude and longitude here as
well as for that parameter means that different sub-sets can be plotted.
(ie. dots without location names, and location names without dots). A default
set of WA towns is included in \code{towns_wa_grainbelt}}

\item{plot_type}{Type of plot as per \code{weather_plot}}

\item{name}{Name of file for output image. Not required if plot.type = 0;
NULL as default}

\item{html_path}{Path for output image file. Not required if plot.type = 0;
NULL as default}

\item{mask_agregion}{Object indicating the mapping region of interest
(non-coastal lines). (optional)}

\item{mask_coast}{Object indicating the location of the ocean in the mapping
area of interest. This is a mask for the mapping, so that nothing is
plotted here. (optional)}

\item{lines_shire}{Object containing the location of the shire boundaries of
the mapping area of interest. This adds lines to the map (optional)}

\item{lines_coast}{Plot object to add the coast line to the map. Default is
coast line for SW WA (optional)}

\item{lines_agregion}{Plot object to add a line indicating the
northern/eastern edge of the grainbelt. (optional)}

\item{logo}{image object for the logo to be added to the map (optional).}

\item{scale}{Plotting scale, default = 10. Changes the resolution of the
saved file, not the relative sizes of the objects}

\item{lambda}{Smoothing parameter for the Kriging. Default is 0.01. Larger
values (e.g. 0.5) will give greater smoothing.}

\item{legendx}{left edge of a vertical legend to be placed using
\code{embed_legend}. Will only plot a legend if this is a numeric
parameter. Should not throw errors.}

\item{transparency}{Should the background be transparent? If false, the
background will be white. (default = FALSE)}

\item{box}{Should a box be drawn around the plot area? (default = TRUE)}

\item{type}{What kind of map are you after? Current options are \code{krig} and
'point', future plans will include options for categorical data (either
points or interpolated). \code{krig} takes a continuous variable and interpolates
to fill the map space with categorical colour; point takes a continuous
variable and plots just the points with categorical colour. Default value
is 'krig' for backwards compatibility}
}
\description{
Generic function for plotting South West WA SSF-style maps. Plots a spatial
map, ensuring that the x and y axes are scaled equally. Plots coloured
rectangles so that plot area is properly filled.
}
\keyword{hplot}
