% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{map_weather_swld}
\alias{map_weather_swld}
\title{Map the SWLD}
\usage{
map_weather_swld(data, varname, col_df, ...)
}
\arguments{
\item{data}{Data frame with columns for the latitude and longitude locations
of the points, and the measurement of \code{varname} at those points. Other
columns will be ignored.}

\item{varname}{Name of variable to map}

\item{col_df}{Data frame containing discrete colour information. For
\code{map_ssf_dots} the defaults is NULL. When this is left as NULL, then a
rainbow of colours is applied to the unique values. Note that this may give
strange results if the variable to be plotted is a factor and the factor
order is not in the desired legend order.}

\item{...}{other parameters as per \code{map_weather} which will be passed to
that function. This excludes the masks and lines, which are hard coded in
this function}
}
\description{
Map the weather for the South West Land Division of Western Australia
}
