% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metno_forecast.R
\name{metno_get_dominant_symbol}
\alias{metno_get_dominant_symbol}
\title{Get dominant weather symbol from a collection}
\usage{
metno_get_dominant_symbol(symbols)
}
\arguments{
\item{symbols}{Character vector of weather symbols}
}
\value{
The dominant weather symbol
}
\description{
Get dominant weather symbol from a collection
}
\seealso{
Other METNO: 
\code{\link{get_metno_daily_forecast}()},
\code{\link{get_metno_forecast}()},
\code{\link{metno_resample_data_table}()},
\code{\link{metno_timeseries_to_data_table}()}

Other parse: 
\code{\link{metno_resample_data_table}()},
\code{\link{metno_timeseries_to_data_table}()},
\code{\link{parse_ag_bulletin}()},
\code{\link{parse_coastal_forecast}()},
\code{\link{parse_precis_forecast}()}
}
\author{
Rodrigo Pires, \email{rodrigo.pires@dpird.wa.gov.au}
}
\concept{METNO}
\concept{parse}
\keyword{internal}
