% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metno_forecast.R
\name{metno_resample_data_table}
\alias{metno_resample_data_table}
\title{Resample data.table to different time frequencies}
\usage{
metno_resample_data_table(dt_data, freq)
}
\arguments{
\item{dt_data}{data.table with time column}

\item{freq}{Character frequency: "hourly", "daily", "weekly", or "monthly"}
}
\value{
A resampled data.table
}
\description{
Resample data.table to different time frequencies
}
\seealso{
Other METNO: 
\code{\link{get_metno_daily_forecast}()},
\code{\link{get_metno_forecast}()},
\code{\link{metno_get_dominant_symbol}()},
\code{\link{metno_timeseries_to_data_table}()}

Other parse: 
\code{\link{metno_get_dominant_symbol}()},
\code{\link{metno_timeseries_to_data_table}()},
\code{\link{parse_ag_bulletin}()},
\code{\link{parse_coastal_forecast}()},
\code{\link{parse_precis_forecast}()}
}
\author{
Rodrigo Pires, \email{rodrigo.pires@dpird.wa.gov.au}
}
\concept{METNO}
\concept{parse}
\keyword{internal}
