% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metno_forecast.R
\name{metno_timeseries_to_data_table}
\alias{metno_timeseries_to_data_table}
\title{Convert MET Weather API timeseries to data.table}
\usage{
metno_timeseries_to_data_table(timeseries)
}
\arguments{
\item{timeseries}{List of timeseries data from MET Weather API}
}
\value{
A data.table with hourly forecast data
}
\description{
Convert MET Weather API timeseries to data.table
}
\seealso{
Other METNO: 
\code{\link{get_metno_daily_forecast}()},
\code{\link{get_metno_forecast}()},
\code{\link{metno_get_dominant_symbol}()},
\code{\link{metno_resample_data_table}()}

Other parse: 
\code{\link{metno_get_dominant_symbol}()},
\code{\link{metno_resample_data_table}()},
\code{\link{parse_ag_bulletin}()},
\code{\link{parse_coastal_forecast}()},
\code{\link{parse_precis_forecast}()}
}
\author{
Rodrigo Pires, \email{rodrigo.pires@dpird.wa.gov.au}
}
\concept{METNO}
\concept{parse}
\keyword{internal}
