% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{parse_silo}
\alias{parse_silo}
\title{SILO API query parser}
\usage{
parse_silo(query_response, this_format, this_date)
}
\arguments{
\item{query_response}{data returned by the SILO API, usually a list of
character data.}

\item{this_format}{A string, user defined by the query details. One of
'alldata', 'apsim' and 'monthly'. Internally inherited from
\code{get_silo_points()}.}

\item{this_date}{A string, user defined by the query details and represents
the start date of the query. Internally inherited from \code{get_silo_points()}.}
}
\value{
A data.frame with date class column(s) and numeric class columns for
the weather variables.
}
\description{
This function takes results from an API query to the SILO
database and formats it to a flat data.frame or a list of data.frames if
querying multiple sites. The function also converts character columns to date
and numerical classes, according to the data represented in the column.
Note that for the 'monthly' \code{data_format} the function also renames the
columns from "Yr.Mth", "Avg TMax (oC)", "Avg TMin (oC)", "Tot Rain (mm)"
"Tot Evap (mm)", "Avg Rad (MJ/m2)", "Avg VP (hPa)" to "year_month", "
tmax_avg", "tmin_avg", "total_rainfall", "total_evap", "radiation_avg" and
"vapour_pressure_avg", respectively.
}
\keyword{internal}
