% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_satellite_imagery.R
\name{plot,SpatRaster,ANY-method}
\alias{plot,SpatRaster,ANY-method}
\alias{plot}
\title{Make a Map of BOM Satellite Data}
\usage{
\S4method{plot}{SpatRaster,ANY}(x, y = "missing", ...)
}
\arguments{
\item{x}{An object of class \code{SpatRaster}.}

\item{y}{Not used.}

\item{...}{Plot parameters forwarded.}
}
\description{
Native plotting of \CRANpkg{terra} \code{SpatRaster} objects, such as objects
returned from \code{get_satellite_imagery()} by default.
}
