% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{write_apsim_met}
\alias{plot}
\title{Objects exported from other packages}
\seealso{
Other APSIM: 
\code{\link{get_data_drill_apsim}()},
\code{\link{get_dpird_apsim}()},
\code{\link{get_patched_point_apsim}()}
}
\concept{APSIM}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{apsimx}{\code{\link[apsimx]{write_apsim_met}}}

  \item{terra}{\code{\link[terra]{plot}}}
}}

