% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shires.lines}
\alias{shires.lines}
\title{South west shire boundaries}
\format{
...
}
\source{
copied from the SSF: unknown -- originally sourced or created by Fiona Evans.
}
\usage{
shires.lines
}
\description{
Set of line segments that make up the south west WA shire boundaries
** current version has bits that go outside the agricultural region
** format needs sorting out
}
\keyword{datasets}
