% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silo_daily_values.R
\docType{data}
\name{silo_daily_values}
\alias{silo_daily_values}
\title{Daily weather data values available from the SILO API}
\format{
A named \code{character} \link{vector} object of 18 items.
}
\source{
\url{https://www.agric.wa.gov.au/weather-api-20}
}
\usage{
silo_daily_values
}
\description{
A \link{vector} object containing 18 named items representing valid values to
supply to \code{get_patched_point()}'s or \code{get_data_drill()}'s \code{values} argument.
The names of the items in the vector correspond to valid values to be
passed along through the \code{which_values} parameter.  The values are what are
passed along to the \acronym{API} query.
}
\keyword{datasets}
