% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_names.R
\name{standardise_names}
\alias{standardise_names}
\title{Standardise Column Names and Row Bind DPIRD and SILO Data}
\usage{
standardise_names(x)
}
\arguments{
\item{x}{An object of S3 class \sQuote{weatherOz.daily.wth} containing daily
weather data returned from \code{get_dpird_summary()}, \code{get_silo_data_drill()}
or \code{get_silo_patched_point()}.}
}
\description{
Takes \emph{daily} \acronym{DPIRD} or \acronym{SILO} weather data returned by
\code{get_dpird_summary()}, \code{get_silo_data_drill()} or
\code{get_silo_patched_point()} and standardises the column names and if
multiple stations are provided, performs \code{rbind()} returning a unified
\CRANpkg{data.table} object.  All columns of the original data are
retained if they are not present between the APIs, \emph{i.e.}, only those data
that have corresponding values in each \acronym{API} are renamed, all
others are kept as-is or filled with \code{NA}.
}
