% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep_for_forecast_towns.R
\name{sweep_for_forecast_towns}
\alias{sweep_for_forecast_towns}
\title{Find nearest BOM forecast towns for a given location}
\usage{
sweep_for_forecast_towns(latlon = c(-35.3, 149.2))
}
\arguments{
\item{latlon}{A length-2 \code{numeric} vector of latitude and longitude in
decimal degree (DD) format. By default, Canberra (approximately).}
}
\value{
A \code{\link{data.table}} of all forecast towns (in this package)
sorted by distance from \var{latlon}, ascending.
}
\description{
Find nearest BOM forecast towns for a given location
}
\examples{

# sweep for forecast towns near Esperance, WA
sweep_for_forecast_towns(latlon = c(-33.8614, 121.8913))

}
\author{
Hugh Parsonage, \email{hugh.parsonage@gmail.com} and
James Goldie, \email{me@jamesgoldie.dev}
}
\concept{bomrang-ported-ported}
