% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_weather_summary.R
\name{tidy_weather_summary}
\alias{tidy_weather_summary}
\title{Individual station summaries nicely formatted.}
\usage{
tidy_weather_summary(
  site = NULL,
  first,
  last = Sys.Date(),
  api_key = NULL,
  interval = c("daily", "15min", "30min", "hourly", "monthly", "yearly"),
  api_name = "weather",
  api_version = 2,
  which_vars = c("all", "rain", "wind", "temp", "erosion")
)
}
\arguments{
\item{site}{A string of the station ID code for the station of interest.
Passed through from \code{get_summaries} function.}

\item{first}{The date on which the weather data summary should start.}

\item{last}{The date on which the weather data summary should end.}

\item{api_key}{Api key from DPIRD (https://www.agric.wa.gov.au/web-apis).
Defaults to NULL.}

\item{interval}{Time interval to summarise over.
Default is 'daily'; others are '15min', '30min', 'hourly',
'monthly', 'yearly'.For intervals shorter than 1 day, time period covered
will be midnight to midnight, with the last time interval being before
midnight - hour/minute values are for the end of the time period.
Data for shorter intervals ('15min', '30min') should be available from
January of last year}

\item{api_name}{Defaults to "weather", Only works with DPIRD's Weather API.}

\item{api_version}{Defaults to 2, and gives a error if not 2.}

\item{which_vars}{Match weather summary selected. Defaults to "all".
Can be one of "all", "rain", "wind", "temp" and "erosion."}
}
\value{
a \verb{data frame} with site and date interval queried together with
requested weather summary/summaries.
}
\description{
Individual station summaries nicely formatted.
}
\examples{
# You must have an DPIRD API key to proceed
mykey <- 'dpird_api_key'

# Set date interval for yearly request
# Get rainfall summary
start.date <- "2017-10-28"

# Use default for end data (current system date)
output <- tidy_weather_summary(
            site = "CL001",
            first = start.date,
            api_key = mykey,
            interval = "yearly",
            which_vars = "rain")

# Only for wind and erosion conditions for daily time interval
# define start and end date
start.date <- "2022-05-01"
end.date <- "2022-05-02"

output <- tidy_weather_summary(
            site = "BI",
            first = start.date,
            last = end.date,
            api_key = mykey,
            interval = "daily",
            which_vars = c("wind", "erosion"))

}
