% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{towns.wa.grainbelt}
\alias{towns.wa.grainbelt}
\title{WA grainbelt towns}
\format{
...
}
\source{
copied from the SSF, from the grainbelt stations list.
}
\usage{
towns.wa.grainbelt
}
\description{
Default set of location names to be added to grainbelt region maps
}
\keyword{datasets}
