% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ag_station_locations.R
\name{update_ag_station_locations}
\alias{update_ag_station_locations}
\title{Update internal databases with latest BOM ag bulletin location data}
\usage{
update_ag_station_locations()
}
\value{
Updated internal databases of \acronym{BOM} station locations and
\acronym{JSON} \acronym{URL}s

\code{NULL}, called for it's side-effects writing an updated file to local
disk for the package to use.
}
\description{
Download the latest ag bulletin station locations and metadata and update
the internal database that supports the use of \code{\link[=get_ag_bulletin]{get_ag_bulletin()}}.
There is no need to use this unless you know that a station exists in
\acronym{BOM}'s database that is not available in the databases distributed
with \pkg{weatherOz}. In fact, for reproducibility purposes. In fact, users
are strongly discouraged from using this function.  Ported from
\pkg{bomrang}.
}
\details{
If \CRANpkg{ASGS.foyer} is installed locally, this function will
automatically check and correct any invalid state values for stations located
in Australia.  If \CRANpkg{ASGS.foyer} is not installed, the function will
update the internal database without validating the state values for stations
by reported longitude/latitude location.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

update_station_ag_locations()
\dontshow{\}) # examplesIf}
}
\references{
Station location and other metadata are sourced from the Australian Bureau of
Meteorology (\acronym{BOM}) webpage, Bureau of Meteorology Site Numbers:\cr
\url{http://www.bom.gov.au/climate/cdo/about/site-num.shtml}.
}
\seealso{
Other bomrang-ported: 
\code{\link{find_forecast_towns}()},
\code{\link{get_ag_bulletin}()},
\code{\link{get_available_imagery}()},
\code{\link{get_available_radar}()},
\code{\link{get_coastal_forecast}()},
\code{\link{get_precis_forecast}()},
\code{\link{get_satellite_imagery}()},
\code{\link{parse_coastal_forecast}()},
\code{\link{parse_precis_forecast}()},
\code{\link{update_forecast_towns}()}
}
\author{
Adam H. Sparks, \email{adam.sparks@dpird.wa.gov.au}
}
\concept{bomrang-ported}
