% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wa_populous_towns}
\alias{wa_populous_towns}
\title{List of WA towns by population}
\format{
...
}
\source{
create by scrapping list of most populous places in WA from Wikipedia
(https://en.wikipedia.org/wiki/List_of_places_in_Western_Australia_by_population).
}
\usage{
wa_populous_towns
}
\description{
Default set locations to be added region maps
}
\keyword{datasets}
