% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries_silo.R
\name{weather.data}
\alias{weather.data}
\alias{bom_data_full}
\title{Weather data}
\usage{
weather.data()

bom_data_full(
  site,
  first,
  last = Sys.Date(),
  email = NULL,
  vars_string = NULL,
  interval = "daily"
)
}
\arguments{
\item{site}{...}

\item{first}{...}

\item{last}{...}

\item{email}{...}

\item{vars_string}{...}

\item{interval}{...}
}
\description{
Fill in later
}
\section{Functions}{
\itemize{
\item \code{bom_data_full}: Daily weather data from BOM owned weather stations,
\code{bom.data.full} returns all of the provided data from the SILO data
set. \code{Date} is returned in date format, \code{Date2} is returned as
character, and the remaining variables are returned as numeric.
}}

