% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{weather_plot}
\alias{weather_plot}
\title{Opens a plot device}
\usage{
weather_plot(
  type = 0,
  width,
  height,
  name = NULL,
  path = NULL,
  background = "white"
)
}
\arguments{
\item{type}{plot type (0 = separate window, 1 = jpg, 4 = png). Default = 0.
Note: png files will give a clearer plot and a smaller file for the same
width/height parameters.}

\item{width}{plot width. These are converted internally; Finished measurement
in inches is ~ width/10. Saved image files are slightly different from
those opened in a new window. This is not a change to be dealt with now.}

\item{height}{plot height (units as per width)}

\item{name}{(something something file name; will have _probs and then file
type appended) Defaults to NULL to allow for opening an unnamed window.}

\item{path}{directory in which the file will be saved for types 1 through 4.
Default is NULL.}

\item{background}{background colour. Will only be relevant if type is 4.}
}
\description{
Opens the specified type of device to fit a plot of the desired aspect ratio
and size. different plot devices require different values to get the same
sized plot. This has been specifically set up for the SWLD of WA, as
per the maps from the SSF. It is included in this package as an interim
function and is likely to be deprecated and removed in later versions.
}
