% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dpird_extreme_weather_values}
\alias{dpird_extreme_weather_values}
\title{A List of DPIRD Extreme Weather Data Values}
\format{
A \link{vector} object of 57 items.
}
\source{
\url{https://www.dpird.wa.gov.au/online-tools/apis/}
}
\usage{
dpird_extreme_weather_values
}
\description{
A \link{vector} object containing 57 items representing valid values to supply
to \code{get_dpird_extremes()}'s \var{values} argument taken from the
documentation for the \acronym{DPIRD} Weather 2.0 \acronym{API}.
}
\seealso{
Other DPIRD: 
\code{\link{dpird_minute_values}},
\code{\link{dpird_summary_values}},
\code{\link{find_nearby_stations}()},
\code{\link{find_stations_in}()},
\code{\link{get_dpird_apsim}()},
\code{\link{get_dpird_availability}()},
\code{\link{get_dpird_extremes}()},
\code{\link{get_dpird_minute}()},
\code{\link{get_dpird_summaries}()},
\code{\link{get_stations_metadata}()}

Other data: 
\code{\link{dpird_minute_values}},
\code{\link{dpird_summary_values}},
\code{\link{silo_daily_values}}
}
\concept{DPIRD}
\concept{data}
\keyword{datasets}
