% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dpird_summary_values}
\alias{dpird_summary_values}
\title{A List of DPIRD Summary Weather Data Values}
\format{
A \link{vector} object of 75 items.
}
\source{
\url{https://www.agric.wa.gov.au/weather-api-20}
}
\usage{
dpird_summary_values
}
\description{
A \link{vector} object containing 75 items representing valid values to supply
to \code{get_dpird_summary()}'s \var{values} argument taken from the
documentation for the \acronym{DPIRD} Weather 2.0 \acronym{API}.
}
\seealso{
Other DPIRD: 
\code{\link{dpird_extreme_weather_values}},
\code{\link{dpird_minute_values}},
\code{\link{find_nearby_stations}()},
\code{\link{find_stations_in}()},
\code{\link{get_dpird_apsim}()},
\code{\link{get_dpird_availability}()},
\code{\link{get_dpird_extremes}()},
\code{\link{get_dpird_minute}()},
\code{\link{get_dpird_summaries}()},
\code{\link{get_stations_metadata}()}

Other data: 
\code{\link{dpird_extreme_weather_values}},
\code{\link{dpird_minute_values}},
\code{\link{silo_daily_values}}
}
\concept{DPIRD}
\concept{data}
\keyword{datasets}
