% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_forecast_towns.R
\name{find_forecast_towns}
\alias{find_forecast_towns}
\title{Find nearest BOM forecast towns for a given location by lat and lon}
\usage{
find_forecast_towns(latitude = -35.3, longitude = 149.2, distance_km = 100)
}
\arguments{
\item{latitude}{A \code{numeric} value of latitude in decimal degree (DD)
format. By default, Canberra (approximately).}

\item{longitude}{A \code{numeric} value of longitude in decimal degree (DD)
format. By default, Canberra (approximately).}

\item{distance_km}{A \code{numeric} value of the distance in kilometres from the
\code{latitude} and \code{longitude} point beyond which values will not be returned.}
}
\value{
A \link[data.table:data.table]{data.table::data.table} of all forecast towns (in this package)
sorted by distance from \var{latlon}, ascending.
}
\description{
Find nearest BOM forecast towns for a given location by lat and lon
}
\examples{

# find forecast towns near Esperance, WA
find_forecast_towns(latitude = -33.8614, longitude = 121.8913)

}
\seealso{
Other bomrang-ported: 
\code{\link{get_ag_bulletin}()},
\code{\link{get_available_imagery}()},
\code{\link{get_available_radar}()},
\code{\link{get_coastal_forecast}()},
\code{\link{get_precis_forecast}()},
\code{\link{get_satellite_imagery}()},
\code{\link{parse_coastal_forecast}()},
\code{\link{parse_precis_forecast}()},
\code{\link{update_forecast_towns}()}
}
\author{
Hugh Parsonage, \email{hugh.parsonage@gmail.com} and
James Goldie, \email{me@jamesgoldie.dev} and Adam Sparks,
\email{adam.sparks@dpird.wa.gov.au}
}
\concept{bomrang-ported}
