% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nearby_stations.R
\name{find_nearby_stations}
\alias{find_nearby_stations}
\title{Get a list of nearby weather stations from coordinates or station ID code}
\usage{
find_nearby_stations(
  latitude = NULL,
  longitude = NULL,
  site = NULL,
  distance_km = 100,
  api_key = NULL,
  wa_only = TRUE,
  dpird_only = FALSE
)
}
\arguments{
\item{latitude}{A \code{numeric} value for latitude expressed as decimal degrees
(DD) (WGS84).}

\item{longitude}{A \code{numeric} value for longitude expressed as decimal degrees
(DD) (WGS84).}

\item{site}{A \code{string} with the station ID code for the station of interest.
Optional and defaults to \code{NULL}.}

\item{distance_km}{A \code{numeric} value for distance to limit the search from
the station or location of interest.  Defaults to 100 km.}

\item{api_key}{A \code{string} value that is the user's \acronym{API} key from
\acronym{DPIRD} (see \url{https://www.agric.wa.gov.au/web-apis}).}

\item{wa_only}{\code{Boolean}, return stations in Western Australia only. Defaults
to \code{TRUE}.}

\item{dpird_only}{\code{Boolean}, return only \acronym{DPIRD} owned stations.
Defaults to \code{FALSE}.}
}
\value{
a \code{data.table} with 'stationCode', 'stationName', 'latitude',
'longitude', 'state', 'owner' and 'distance'. Data is sorted by increasing
distance from station of interest.
}
\description{
Get a list of nearby weather stations from coordinates or station ID code
}
\examples{
# You must have an DPIRD API key to proceed
my_key <- rstudioapi::askForSecret()

# Query WA only stations and return both DPIRD's and BOM's stations for
# the Northam WA station, returning stations with 50 km of this station
wa_stn <- get_nearby_stations(
  site = "NO",
  distance_km = 50,
  api_key = "YOUR API KEY",
  dpird_only = FALSE,
  wa_only = TRUE
)

# Query Wagga Wagga BOM station.
wagga_stn <- get_nearby_stations(
  latitude = -35.1583,
  longitude = 147.4575,
  distance_km = my_distance,
  api_key = "YOUR API KEY",
  dpird_only = FALSE,
  wa_only = FALSE
)

}
\author{
Rodrigo Pires, \email{rodrigo.pires@dpird.wa.gov.au}
}
