% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nearby_stations.R
\name{find_nearby_stations}
\alias{find_nearby_stations}
\title{Get a list of nearby weather stations from coordinates or station ID code}
\usage{
find_nearby_stations(
  latitude = NULL,
  longitude = NULL,
  station_id = NULL,
  distance_km = 100,
  api_key = NULL,
  which_api = "silo"
)
}
\arguments{
\item{latitude}{A \code{numeric} value for latitude expressed as decimal degrees
(DD) (WGS84).}

\item{longitude}{A \code{numeric} value for longitude expressed as decimal degrees
(DD) (WGS84).}

\item{station_id}{A \code{string} with the station ID code for the station of
interest. Optional and defaults to \code{NULL}.}

\item{distance_km}{A \code{numeric} value for distance to limit the search from
the station or location of interest.  Defaults to 100 km.}

\item{api_key}{A \code{string} value that is the user's \acronym{API} key from
\acronym{DPIRD} (see \url{https://www.agric.wa.gov.au/web-apis}).}

\item{silo_stations}{\code{Boolean}, return SILO stations only. Defaults
to \code{TRUE}. If \code{FALSE}, return only \acronym{DPIRD} owned stations.}
}
\value{
a \code{data.table} with 'station_code', 'station_name', 'latitude',
'longitude', 'elevation' and 'distance'. Data are sorted by increasing
distance from station or location of interest. SILO queries also return the
Australian State ('state') while DPIRD queries also returns 'owner', 'online'
and 'status'.
}
\description{
Get a list of nearby weather stations from coordinates or station ID code
}
\note{
You can request your own API key from DPIRD for free by filling out the
form found at \url{https://www.agric.wa.gov.au/web-apis}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# You must have an DPIRD API key to proceed
my_key <- rstudioapi::askForSecret()

# Query WA only stations and return both DPIRD's and BOM's stations for
# the Northam WA station, returning stations with 50 km of this station
wa_stn <- find_nearby_stations(
  station_id = "NO",
  distance_km = 50,
  api_key = "YOUR API KEY",
  silo_stations = FALSE
)

# Query Wagga Wagga BOM station.
wagga_stn <- find_nearby_stations(
  latitude = -35.1583,
  longitude = 147.4575,
  distance_km = 200,
  api_key = "YOUR API KEY",
  silo_stations = FALSE
)
\dontshow{\}) # examplesIf}
}
\author{
Rodrigo Pires, \email{rodrigo.pires@dpird.wa.gov.au}
}
