% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nearby_stations.R
\name{find_nearby_stations}
\alias{find_nearby_stations}
\title{Find nearby weather stations given geographic coordinates or a station code}
\usage{
find_nearby_stations(
  latitude = NULL,
  longitude = NULL,
  station_code = NULL,
  distance_km = 100,
  api_key = NULL,
  which_api = "silo"
)
}
\arguments{
\item{latitude}{A \code{numeric} value for latitude expressed as decimal degrees
(DD) (WGS84).}

\item{longitude}{A \code{numeric} value for longitude expressed as decimal degrees
(DD) (WGS84).}

\item{station_code}{A \code{string} with the station code for the station of
interest. Optional and defaults to \code{NULL}.}

\item{distance_km}{A \code{numeric} value for distance to limit the search from
the station or location of interest.  Defaults to 100 km.}

\item{api_key}{A \code{string} value that is the user's \acronym{API} key from
\acronym{DPIRD} (see \url{https://www.agric.wa.gov.au/web-apis}).  Only used
when \code{which_api} is "DPIRD" or "all".}

\item{which_api}{A \code{string} value that indicates which API to use.  Defaults
to "silo". Valid values are "all", for both \acronym{SILO} (\acronym{BOM})
and \acronym{DPIRD} weather station networks; "silo" for only stations in
the \acronym{SILO} network; or "dpird" for stations in the \acronym{DPIRD}
network.}
}
\value{
a \code{data.table} with 'station_code', 'station_name', 'latitude',
'longitude', 'elevation', 'state', 'owner', and 'distance'.  Data are sorted
by increasing distance from station or location of interest.
}
\description{
Find nearby weather stations given geographic coordinates or a station code
}
\note{
You can request your own \acronym{API} key from \acronym{DPIRD} for
free by filling out the form found at \url{https://www.agric.wa.gov.au/web-apis}.
}
\examples{
\dontrun{
# Query WA only stations and return both DPIRD's and BOM's stations for
# the Northam WA station, returning stations with 50 km of this station
wa_stn <- find_nearby_stations(
  station_code = "NO",
  distance_km = 50,
  api_key = "YOUR API KEY",
  which_api = "DPIRD"
)

# Query Wagga Wagga BOM station.
wagga_stn <- find_nearby_stations(
  latitude = -35.1583,
  longitude = 147.4575,
  distance_km = 200,
  which_api = "SILO"
)
}

}
\author{
Rodrigo Pires, \email{rodrigo.pires@dpird.wa.gov.au}
}
