% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_drill.R
\name{get_data_drill}
\alias{get_data_drill}
\title{Get DataDrill weather data from the SILO API}
\usage{
get_data_drill(
  longitude,
  latitude,
  start_date,
  end_date = Sys.Date(),
  which_values = "all",
  interval = "daily",
  api_key
)
}
\arguments{
\item{longitude}{A single \code{numeric} value  representing the longitude of the
point-of-interest.}

\item{latitude}{A single \code{numeric} value representing the latitude of the
point-of-interest.}

\item{start_date}{A \code{character} string representing the beginning of the
range to query in the format 'yyyy-mm-dd' (ISO8601).  Will return data
inclusive of this range.}

\item{end_date}{A \code{character} string representing the end of the range query
in the format 'yyyy-mm-dd' (ISO8601).  Will return data inclusive of this
range.  Defaults to the current system date.}

\item{which_values}{A \code{character} string with the type of weather data to
return.  See \strong{Available Values} for a full list of valid values.
Defaults to 'all' with all available values being returned.}

\item{interval}{A \code{character} string that indicates the time interval
requested.  Default is "daily", optionally "monthly" and "yearly" are
available.  See Details for more.}

\item{api_key}{A \code{character }string specifying a valid email address to use
for the request.  The query will return an error if a valid email address
is not provided.}
}
\value{
a \link[data.table:data.table]{data.table::data.table} with \code{longitude} and \code{latitude} along with
the date interval queried together with the requested weather variables in
alphabetical order. The first six columns will always be:
\itemize{
\item longitude,
\item latitude,
\item year,
\item month (if daily or monthly)
\item day (if daily), and
\item date (if daily or monthly ISO8601 format, "YYYYMMDD")
}
}
\description{
Download weather data from the \acronym{SILO} \acronym{API} for spatially
interpolated weather data (DataDrill).  The daily climate surfaces have
been derived either by splining or kriging the observational data.  The
grid spans 112° to 154°, -10° to -44° with resolution 0.05° latitude by
0.05° longitude (approximately 5 km × 5 km).
}
\details{
When \code{interval} is a summary, \emph{i.e.}, "monthly" or "yearly", (i) monthly or
yearly totals for rainfall and evaporation; and (ii) monthly or yearly
means for maximum and minimum temperatures, solar radiation and vapour
pressure are returned.
}
\section{Available Values}{


\describe{
\item{all}{Which will return all of the following values}
\item{rain}{Rainfall}
\item{max_temp}{Maximum temperature}
\item{min_temp}{Minimum temperature}
\item{vp}{Vapour pressure}
\item{vp_deficit}{Vapour pressure deficit}
\item{evap_pan}{Class A pan evaporation}
\item{evap_syn}{Synthetic estimate^1}
\item{evap_comb}{Combination (synthetic estimate pre-1970, class A pan 1970
onwards)}
\item{evap_morton_lake}{Morton's shallow lake evaporation}
\item{radiation}{Solar exposure, consisting of both direct and diffuse
components}
\item{rh_tmax}{Relative humidity at the time of maximum temperature}
\item{rh_tmin}{Relative humidity at the time of minimum temperature}
\item{et_short_crop}{FAO56^4 short crop}
\item{et_tall_crop}{ASCE^5 tall crop^6}
\item{et_morton_actual}{Morton's areal actual evapotranspiration}
\item{et_morton_potential}{Morton's point potential evapotranspiration}
\item{et_morton_wet}{Morton's wet-environment areal potential
evapotranspiration over land}
\item{mslp}{Mean sea level pressure}
}
}

\section{Value information}{


Solar radiation: total incoming downward shortwave radiation on a horizontal
surface, derived from estimates of cloud oktas and sunshine duration^3.

Relative humidity: calculated using the vapour pressure measured at 9am, and
the saturation vapour pressure computed using either the maximum or minimum
temperature^6.

Evaporation and evapotranspiration: an overview of the variables provided by
\acronym{SILO} is available here,
\url{https://data.longpaddock.qld.gov.au/static/publications/Evapotranspiration_overview.pdf}.
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

#' # Source data from latitude and longitude coordinates (gridded data)
# Southwood, QLD.
wd <- get_silo(latitude = -27.85,
               longitude = 150.05,
               start_date = "20221001",
               end_date = "20221201",
               api_key = "your@email")
\dontshow{\}) # examplesIf}
}
\references{
\enumerate{
\item Rayner, D. (2005). Australian synthetic daily Class A pan evaporation.
Technical Report December 2005, Queensland Department of Natural Resources
and Mines, Indooroopilly, Qld., Australia, 40 pp.
\item Morton, F. I. (1983). Operational estimates of areal evapotranspiration
and their significance to the science and practice of hydrology, \emph{Journal
of Hydrology}, Volume 66, 1-76.
\item Zajaczkowski, J., Wong, K., & Carter, J. (2013). Improved historical
solar radiation gridded data for Australia, \emph{Environmental Modelling &
Software}, Volume 49, 64–77. DOI: \doi{10.1016/j.envsoft.2013.06.013}.
\item Food and Agriculture Organization of the United Nations,
Irrigation and drainage paper 56: Crop evapotranspiration - Guidelines for
computing crop water requirements, 1998.
\item ASCE’s Standardized Reference Evapotranspiration Equation, proceedings of
the National Irrigation Symposium, Phoenix, Arizona, 2000.
\item For further details refer to Jeffrey, S.J., Carter, J.O., Moodie, K.B. and
Beswick, A.R. (2001). Using spatial interpolation to construct a
comprehensive archive of Australian climate data, \emph{Environmental Modelling
and Software}, Volume 16/4, 309-330. DOI:
\doi{10.1016/S1364-8152(01)00008-1}.
}
}
\seealso{
Other SILO: 
\code{\link{get_patched_point}()}
}
\concept{SILO}
