% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dpird_apsim.R
\encoding{UTF-8}
\name{get_dpird_apsim}
\alias{get_dpird_apsim}
\title{Get DPIRD Summary Weather Data in the APSIM Format From the Weather 2.0 API}
\usage{
get_dpird_apsim(station_code, start_date, end_date = Sys.Date(), api_key)
}
\arguments{
\item{station_code}{A \code{character} string of the \acronym{DPIRD} station code
for the station of interest.  Station codes are available from the
\code{get_stations_metadata()} function.}

\item{start_date}{A \code{character} string or \code{Date} object representing the
beginning of the range to query in the format \dQuote{yyyy-mm-dd}
(ISO8601).  Data returned is inclusive of this date.}

\item{end_date}{A \code{character} string or \code{Date} object representing the end of
the range query in the format  \dQuote{yyyy-mm-dd} (ISO8601).  Data
returned is inclusive of this date.  Defaults to the current system date.}

\item{api_key}{A \code{character} string containing your \acronym{API} key from
\acronym{DPIRD}, \url{https://www.agric.wa.gov.au/web-apis}, for the
\acronym{DPIRD} Weather 2.0 \acronym{API}.}
}
\value{
An \CRANpkg{apsimx} object of class \sQuote{met} with attributes,
alternatively, a local APSIM .met file if \code{file} is specified.
}
\description{
Automates the retrieval and conversion of summary data from the
\acronym{DPIRD} Weather 2.0 \acronym{API} to an \acronym{APSIM} .met file
formatted weather data object.
}
\section{Saving objects}{

To save \dQuote{met} objects, please use \code{\link[apsimx:write_apsim_met]{apsimx::write_apsim_met()}}.
}

\examples{
\dontrun{
# Get an APSIM format object for Binnu
# Note that you need to supply your own API key

wd <- get_dpird_apsim(
  station_code = "BI",
  start_date = "20220401",
  end_date = "20221101",
  api_key = "your_api_key"
)
}


}
\seealso{
Other DPIRD: 
\code{\link{dpird_extreme_weather_values}},
\code{\link{dpird_minute_values}},
\code{\link{dpird_summary_values}},
\code{\link{find_nearby_stations}()},
\code{\link{get_dpird_availability}()},
\code{\link{get_dpird_extremes}()},
\code{\link{get_dpird_minute}()},
\code{\link{get_dpird_summaries}()},
\code{\link{get_stations_metadata}()}

Other data fetching: 
\code{\link{get_ag_bulletin}()},
\code{\link{get_coastal_forecast}()},
\code{\link{get_data_drill}()},
\code{\link{get_data_drill_apsim}()},
\code{\link{get_dpird_extremes}()},
\code{\link{get_dpird_minute}()},
\code{\link{get_dpird_summaries}()},
\code{\link{get_patched_point}()},
\code{\link{get_patched_point_apsim}()},
\code{\link{get_precis_forecast}()},
\code{\link{get_radar_imagery}()},
\code{\link{get_satellite_imagery}()}

Other APSIM: 
\code{\link{get_data_drill_apsim}()},
\code{\link{get_patched_point_apsim}()}
}
\author{
Adam H. Sparks, \email{adamhsparks@gmail.com}
}
\concept{APSIM}
\concept{DPIRD}
\concept{data fetching}
