% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dpird_minute.R
\name{get_dpird_minute}
\alias{get_dpird_minute}
\title{Get minute weather data from DPIRD Weather 2.0 API}
\usage{
get_dpird_minute(
  station_code,
  start_date_time = lubridate::now() - lubridate::hours(24),
  end_date_time = lubridate::now(),
  api_key,
  which_values
)
}
\arguments{
\item{station_code}{A \code{character} string or \code{vector} of the \acronym{DPIRD}
station \acronym{ID} code(s) for the station(s) of interest.}

\item{start_date_time}{A \code{character} string representing the start date and
time of the query in the format 'yyyy-mm-dd-hh-mm'. Defaults to 24 hours
before the current local system time, returning the most recent 24 hour
observations rounded to the nearest minute. This function does its best to
decipher many date and time formats but prefers ISO8601.}

\item{end_date_time}{A \code{character} string representing the start date of the
query in the format 'yyyy-mm-dd-hh-mm'.  Defaults to the current system
date rounded to the nearest minute.  This function does its best to
decipher many date and time formats but prefers ISO8601.}

\item{api_key}{A \code{character} string containing your \acronym{API} key from
\acronym{DPIRD}, see \url{https://www.agric.wa.gov.au/web-apis} for the
\acronym{DPIRD} Weather 2.0 \acronym{API}.}

\item{which_values}{A \code{vector} of weather values to query from the
\acronym{API}. See \strong{Available Values} section for valid available codes.
Defaults to all available values.}
}
\value{
a \code{data.table} with 'station_code' and date interval queried together
with the requested weather variables.
}
\description{
Nicely formatted minute weather station data from the \acronym{DPIRD} weather
station network over a maximum 24 hour period.
}
\note{
Please note this function converts date-time columns from Coordinated
Universal Time 'UTC' returned by the \acronym{API} to Australian Western
Standard Time 'AWST'.
}
\section{Available Values}{

\itemize{
\item airTemperature,
\item dateTime,
\item dewPoint,
\item rainfall,
\item relativeHumidity,
\item soilTemperature,
\item solarIrradiance,
\item wetBulb,
\item wind,
\item windAvgSpeed,
\item windMaxSpeed, and
\item windMinSpeed
}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(lubridate)

today <- now()
yesterday <- now() - hours(24)

get_dpird_minute(station_code = "NO",
                 start_date_time = yesterday,
                 end_date_time = today,
                 api_key = YOUR_API_KEY,
                 which_values = c("airTemperature", "solarIrradiance"))
\dontshow{\}) # examplesIf}
}
\seealso{
Other DPIRD: 
\code{\link{get_dpird_summaries}()},
\code{\link{get_extreme_weather}()}
}
\concept{DPIRD}
