% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dpird_summaries.R
\name{get_dpird_summaries}
\alias{get_dpird_summaries}
\title{Get weather data from DPIRD weather API summarised by time interval}
\usage{
get_dpird_summaries(
  station_id = NULL,
  first = NULL,
  last = Sys.Date(),
  api_key = NULL,
  interval = "daily",
  which_vars = "all"
)
}
\arguments{
\item{station_id}{A string of the station ID code for the station of
interest. Defaults to NULL.}

\item{first}{A string representing the start date of the query in the
format 'yyyy-mm-dd'. Defaults to NULL.}

\item{last}{A string representing the start date of the query in the
format 'yyyy-mm-dd'. Defaults to the current system date.}

\item{api_key}{Api key from DPIRD (https://www.agric.wa.gov.au/web-apis).
Defaults to NULL.}

\item{interval}{Time interval to summarise over.
Default is 'daily'; others are '15min', '30min', 'hourly',
'monthly', 'yearly'.For intervals shorter than 1 day, time period covered
will be midnight to midnight, with the last time interval being before
midnight - hour/minute values are for the end of the time period.
Data for shorter intervals ('15min', '30min') are available from
January of the previous year.}

\item{which_vars}{Match weather summary selected. Defaults to "all".
Can be one of "all", "rain", "wind", "temp" and "erosion."}
}
\value{
a \verb{data table} with station_id and date interval queried together
with the requested weather variables.
}
\description{
Nicely formatted individual station weather summaries from the DPIRD
weather station network.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# You must have an DPIRD API key to proceed
mykey <- 'dpird_api_key'

# Set date interval for yearly request
# Get rainfall summary
start_date <- "2017-10-28"

# Use default for end data (current system date)
output <- get_dpird_summaries(
            station_id = "CL001",
            first = start_date,
            api_key = mykey,
            interval = "yearly",
            which_vars = "rain")

# Only for wind and erosion conditions for daily time interval
# define start and end date
start_date <- "2022-05-01"
end_date <- "2022-05-02"

output <- get_dpird_summaries(
            station_id = "BI",
            first = start_date,
            last = end_date,
            api_key = mykey,
            interval = "daily",
            which_vars = c("wind", "erosion"))
\dontshow{\}) # examplesIf}
}
\seealso{
Other DPIRD: 
\code{\link{get_extreme_weather}()}
}
\concept{DPIRD}
