% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dpird_summaries.R
\name{get_dpird_summaries}
\alias{get_dpird_summaries}
\title{Get weather data from DPIRD Weather 2.0 API summarised by time interval}
\usage{
get_dpird_summaries(
  station_code,
  first,
  last = Sys.Date(),
  api_key,
  interval = "daily",
  which_vars = "all"
)
}
\arguments{
\item{station_code}{A \code{character} string or \code{vector} of the \acronym{DPIRD}
station \acronym{ID} code(s) for the station(s) of interest.}

\item{first}{A \code{character} string representing the start date of the query in
the format 'yyyymmdd'.}

\item{last}{A \code{character} string representing the start date of the query in
the format 'yyyymmdd'.  Defaults to the current system date.}

\item{api_key}{A \code{character} string containing your \acronym{API} key from
\acronym{DPIRD}, \url{https://www.agric.wa.gov.au/web-apis}, for the
\acronym{DPIRD} weather \acronym{API}.}

\item{interval}{A \code{character} string that indicates the time interval to
summarise over.  Default is 'daily'; others are '15min', '30min', 'hourly',
'monthly' or 'yearly'.  For intervals shorter than 1 day, the time period
covered will be midnight to midnight, with the last time interval being
before midnight - hour/minute values are for the end of the time period.
Data for shorter intervals ('15min', '30min') are available from January of
the previous year.}

\item{which_vars}{A \code{character} string selecting the desired weather summary.
Defaults to 'all'.  Can be one of 'all', 'rain', 'wind', 'temp', 'soil' or
'erosion'.}
}
\value{
a \code{data.table} with 'station_code' and date interval queried together
with the requested weather variables.
}
\description{
Nicely formatted individual station weather summaries from the
\acronym{DPIRD} weather station network.
}
\note{
Please note this function converts date-time columns from Coordinated
Universal Time 'UTC' to Australian Western Standard Time 'AWST'.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# You must have an DPIRD API key to proceed
# Set date interval for yearly request
# Get rainfall summary
start_date <- "20171028"

# Use default for end data (current system date)
output <- get_dpird_summaries(
            station_code = "CL001",
            first = start_date,
            api_key = "YOUR API KEY",
            interval = "yearly",
            which_vars = "rain")

# Only for wind and erosion conditions for daily time interval
# define start and end date
start_date <- "20220501"
end_date <- "20220502"

output <- get_dpird_summaries(
            station_code = "BI",
            first = start_date,
            last = end_date,
            api_key = "YOUR API KEY",
            interval = "daily",
            which_vars = c("wind", "erosion"))
\dontshow{\}) # examplesIf}
}
\seealso{
Other DPIRD: 
\code{\link{get_extreme_weather}()}
}
\concept{DPIRD}
