% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dpird_summaries.R
\name{get_dpird_summaries}
\alias{get_dpird_summaries}
\title{Get weather data from DPIRD Weather 2.0 API summarised by time interval}
\usage{
get_dpird_summaries(
  station_code,
  start_date,
  end_date = Sys.Date(),
  interval = "daily",
  which_values = "all",
  api_group = "rtd",
  include_closed = FALSE,
  api_key
)
}
\arguments{
\item{station_code}{A \code{character} string of the \acronym{DPIRD} station code
for the station of interest.}

\item{start_date}{A \code{character} string representing the beginning of the
range to query in the format 'yyyy-mm-dd' (ISO8601).  Will return data
inclusive of this range.}

\item{end_date}{A \code{character} string representing the end of the range query
in the format 'yyyy-mm-dd' (ISO8601).  Will return data inclusive of this
range.  Defaults to the current system date.}

\item{interval}{A \code{character} string that indicates the time interval to
summarise over.  Default is 'daily'; others are '15min', '30min', 'hourly',
'monthly' or 'yearly'.  For intervals shorter than 1 day, the time period
covered will be midnight to midnight, with the end_date time interval being
before midnight - hour/minute values are for the end of the time period.
Data for shorter intervals ('15min', '30min') are available from January of
the previous year.}

\item{which_values}{A \code{character} string with the type of summarised weather
to return.  See \strong{Available Values} for a full list of valid values.
Defaults to 'all' with all available values being returned.}

\item{api_group}{Filter the stations to a predefined group one of 'all',
'web' or 'rtd'; 'all' returns all stations, 'api' returns the default
stations in use with the \acronym{API} and 'web' returns the list in use by
the \url{https:://weather.agric.wa.gov.au} and 'rtd' returns stations with
scientifically complete data sets. Defaults to 'rtd'.}

\item{include_closed}{A \code{Boolean} value that defaults to \code{FALSE}. If set to
\code{TRUE} the query returns closed and open stations. Closed stations are those
that have been turned off and no longer report data. They may be useful for
historical purposes.}

\item{api_key}{A \code{character} string containing your \acronym{API} key from
\acronym{DPIRD}, \url{https://www.agric.wa.gov.au/web-apis}, for the
\acronym{DPIRD} Weather 2.0 \acronym{API}.
\subsection{Available Values for \code{which_values}}{
\itemize{
\item all (which will return all of the following values),
\item airTemperature,
\item airTemperatureAvg,
\item airTemperatureMax,
\item airTemperatureMaxTime,
\item airTemperatureMin,
\item airTemperatureMinTime,
\item apparentAirTemperature,
\item apparentAirTemperatureAvg,
\item apparentAirTemperatureMax,
\item apparentAirTemperatureMaxTime,
\item apparentAirTemperatureMin,
\item apparentAirTemperatureMinTime,
\item barometricPressure,
\item barometricPressureAvg,
\item barometricPressureMax,
\item barometricPressureMaxTime,
\item barometricPressureMin,
\item barometricPressureMinTime,
\item battery,
\item batteryMinVoltage,
\item batteryMinVoltageDateTime,
\item chillHours,
\item deltaT,
\item deltaTAvg,
\item deltaTMax,
\item deltaTMaxTime,
\item deltaTMin,
\item deltaTMinTime,
\item dewPoint,
\item dewPointAvg,
\item dewPointMax,
\item dewPointMaxTime,
\item dewPointMin,
\item dewPointMinTime,
\item erosionCondition,
\item erosionConditionMinutes,
\item erosionConditionStartTime,
\item errors,
\item etoShortCrop,
\item etoTallCrop,
\item evapotranspiration,
\item frostCondition,
\item frostConditionMinutes,
\item frostConditionStartTime,
\item heatCondition,
\item heatConditionMinutes,
\item heatConditionStartTime,
\item observations,
\item observationsCount,
\item observationsPercentage,
\item panEvaporation,
\item rainfall,
\item relativeHumidity,
\item relativeHumidityAvg,
\item relativeHumidityMax,
\item relativeHumidityMaxTime,
\item relativeHumidityMin,
\item relativeHumidityMinTime,
\item richardsonUnits,
\item soilTemperature,
\item soilTemperatureAvg,
\item soilTemperatureMax,
\item soilTemperatureMaxTime,
\item soilTemperatureMin,
\item soilTemperatureMinTime,
\item solarExposure,
\item wetBulb,
\item wetBulbAvg,
\item wetBulbMax,
\item wetBulbMaxTime,
\item wetBulbMin,
\item wetBulbMinTime,
\item wind,
\item windAvgSpeed, and
\item windMaxSpeed
}
}}
}
\value{
a \link[data.table:data.table]{data.table::data.table}  with 'station_code' and date interval
queried together with the requested weather variables in alphabetical order.
The first ten columns will always be:
\itemize{
\item 'station_code',
\item 'station_name',
\item 'period.from',
\item 'period.to',
\item 'period.year',
\item 'period.month',
\item 'period.day',
\item 'period.hour',
\item 'period.minute'
\item 'date' (a combination of year, month, day, hour, minute as appropriate).
}
}
\description{
Nicely formatted individual station weather summaries from the
\acronym{DPIRD} weather station network.
}
\note{
Please note this function converts date-time columns from Coordinated
Universal Time 'UTC' to Australian Western Standard Time 'AWST'.
}
\examples{
\dontrun{
# You must have an DPIRD API key to proceed
# Set date interval for yearly request
# Get rainfall summary
start_date <- "20171028"

# Use default for end data (current system date)
output <- get_dpird_summaries(
            station_code = "CL001",
            start_date = start_date,
            api_key = "YOUR API KEY",
            interval = "yearly",
            which_values = "rainfall")

# Only for wind and erosion conditions for daily time interval
# define start and end date
start_date <- "20220501"
end_date <- "20220502"

output <- get_dpird_summaries(
            station_code = "BI",
            start_date = start_date,
            end_date = end_date,
            api_key = "YOUR API KEY",
            interval = "daily",
            which_values = c("wind", "erosion"))
}
}
\seealso{
Other DPIRD: 
\code{\link{get_dpird_minute}()},
\code{\link{get_extreme_weather}()}
}
\concept{DPIRD}
