% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extreme_weather.R
\name{get_extreme_weather}
\alias{get_extreme_weather}
\title{Get extreme weather event summaries for a single DPIRD station}
\usage{
get_extreme_weather(site, type = "all", api_key = NULL)
}
\arguments{
\item{site}{A string with the station ID code for the station of interest.}

\item{type}{A string with the type of extreme weather to return. Defaults to
"all"; and can be combination of "frost", "erosion", "heat", or "all".}

\item{api_key}{User's \acronym{API} key from \acronym{DPIRD}
(\url{https://www.agric.wa.gov.au/web-apis})}
}
\value{
a \code{data.frame} of one row with 'stationCode', 'stationName',
'latitude', 'longitude', 'dateTime' of the query and the extreme weather
information according to the type(s) selected.
}
\description{
Get extreme weather event summaries for a single DPIRD station
}
\examples{
# You must have an DPIRD API key to proceed
my_key <- rstudioapi::askForSecret()

# Query Bonnie Rock station for wind erosion and heat extreme events.
my_station <- "BR"
my_events <- c("erosion", "heat")

output <- get_extreme_weather(
  site = my_station,
  type = my_events,
  api_key = my_key
)

# Query multiple stations for all extreme events
# Provide a list of station (as strings)
these_stations <- list("MN", "ES", "KARI", "NO", "KA002", "CO001", "MA002")

# Row bind output lists (one for each station)
# together with `data.table::rbindlist`
outputs <- lapply(these_stations,
                  get_extreme_weather,
                  type = "all",
                  api_key = apiKey)

 data.table::rbindlist(outputs)

}
\seealso{
Other DPIRD: 
\code{\link{get_station_list}()},
\code{\link{get_summaries}()},
\code{\link{tidy_weather_summary}()}
}
\author{
Rodrigo Pires, \email{rodrigo.pires@dpird.wa.gov.au}
}
\concept{DPIRD}
