% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_silo.R
\name{get_silo}
\alias{get_silo}
\title{Get weather data from SILO (Scientific Information for Land Owners) API}
\usage{
get_silo(
  station_code = NULL,
  latitude = NULL,
  longitude = NULL,
  start_date,
  end_date = Sys.Date(),
  data_format = "alldata",
  email
)
}
\arguments{
\item{station_code}{An \code{integer} representing the station number
available from the \acronym{SILO} network.  Defaults to \code{NULL} and when
used, queries with latitude and longitude input are not permitted.}

\item{latitude}{A single \code{numeric} value representing the latitude of the
point-of-interest.  Defaults to \code{NULL} and when used, queries with
\code{station_code} input are not permitted. Requires \code{longitude} to be
provided.}

\item{longitude}{A single \code{numeric} value  representing the longitude of the
point-of-interest.  Defaults to \code{NULL} and when used, queries with
\code{station_code} inputs are not permitted.  Requires \code{latitude} to be
provided.}

\item{start_date}{A \code{character} string representing the start date of the
query in the format 'yyyymmdd'.}

\item{end_date}{A \code{character} string representing the start date of the query
in the format 'yyyymmdd'.  Defaults to the current system date.}

\item{data_format}{A \code{character} string specifying the type of data to
retrieve.  Limited to 'alldata', 'monthly' or 'apsim'.  Note 'apsim' and
'alldata' retrieve daily data.}

\item{email}{A \code{character }string specifying a valid email address to use
for the request. The query will return an error if a valid email address is
not provided.}
}
\value{
A \link[data.table:data.table]{data.table::data.table} containing the retrieved data from the
\acronym{SILO} \acronym{API}.
}
\description{
Download weather data from the \acronym{SILO} \acronym{API} from both
station observations (DataDrill) and gridded data (PatchedPointData). There
are three formats available: 'alldata' and 'apsim' with daily frequency and
'monthly' with, that's right, monthly frequency. Queries with
\code{station_code} return stations observations from the 'DataDrill' endpoint
while queries with \code{latitude} and \code{longitude} coordinates return gridded
data from the 'PatchedPointData'.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Source observation data for station Wongan Hills station, WA (8137)
wd <- get_silo(station_code = 8137,
               start_date = "20210601",
               end_date = "20210701",
               data_format = "alldata",
               email = "your@email")

# Source data from latitude and longitude coordinates (gridded data)
# Southwood, QLD in the 'apsim' format.
wd <- get_silo(latitude = -27.85,
               longitude = 150.05,
               start_date = "20221001",
               end_date = "20221201",
               data_format = "apsim",
               email = "your@email")
\dontshow{\}) # examplesIf}
}
\concept{SILO}
