% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_list.R
\name{get_station_list}
\alias{get_station_list}
\title{Fetch a list with all available weather stations for either \code{Science} and/or
\code{Weather} \acronym{DPIRD} \acronym{APIs}.}
\usage{
get_station_list(
  api = c("science", "weather"),
  station_group = c("all", "api", "rtd", "web", "yshistory", "yellowspot"),
  state = c("all", "wa", "sa", "nsw", "vic", "qld", "tas", "nt"),
  api_key = NULL
)
}
\arguments{
\item{api_key}{User's \acronym{API} key from \acronym{DPIRD}
(\url{https://www.agric.wa.gov.au/web-apis})}

\item{which_api}{name of the \acronym{API} to be queried. Needs to match to
one of the existing options (currently 'science' or 'weather')}
}
\value{
a \code{data.frame} with station code, name, longitude, longitude, owner
and state (for the Science API).
Science \acronym{API} queries return additional details by default
\emph{i.e.}, station model, battery voltage, height of sensors, status and others.
}
\description{
Providers include: \acronym{DPIRD}, \acronym{DBCA}, \acronym{DFES},
Water Corporation, Harvey Water and Pardoo Beef Corporation.
}
\examples{
# You must have an DPIRD API key to proceed
mykey <- rstudioapi::askForSecret()

stations_weather <- get_station_list(which_api = "weather",
station_ group = "rtd",
state = "all",
api_key = mykey)

stations_science <- get_station_list(which_api = "science",
station_ group = "yellowspot",
state = "nsw",
api_key = mykey)

}
\author{
Rodrigo Pires, \email{rodrigo.pires@dpird.wa.gov.au}
}
