% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_list.R
\name{get_station_list}
\alias{get_station_list}
\title{Fetch a list with all available weather stations for either \code{Science} and/or
\code{Weather} \acronym{DPIRD} \acronym{APIs}.}
\usage{
get_station_list(
  api = "science",
  station_group = "rtd",
  state = "wa",
  api_key = NULL
)
}
\arguments{
\item{api}{name of the \acronym{API} to be queried. Needs to match to
one of the existing options (currently 'science' or 'weather') and defaults
to the 'science' \acronym{API}.}

\item{station_group}{the name of the station group that you wish to query.
Defaults to 'rtd'. Currently "all", "api", "rtd", "web", "yshistory",
"yellowspot" are valid groups.}

\item{state}{a string limiting the query to one or several states, defaults
to 'wa'. Accepts "all", "wa", "sa", "nsw", "vic", "qld", "tas", "nt".}

\item{api_key}{User's \acronym{API} key from \acronym{DPIRD}
(\url{https://www.agric.wa.gov.au/web-apis})}
}
\value{
a \code{data.frame} with station code, name, latitude, longitude, owner
and state (for the Science API).
Science \acronym{API} queries return additional details by default
\emph{i.e.}, station model, battery voltage, height of sensors, status and others.
}
\description{
Providers include: \acronym{DPIRD}, \acronym{DBCA}, \acronym{DFES},
Water Corporation, Harvey Water and Pardoo Beef Corporation.
}
\examples{
# You must have an DPIRD API key to proceed
mykey <- rstudioapi::askForSecret()

stations_weather <- get_station_list(which_api = "weather",
station_group = "rtd",
state = "all",
api_key = mykey)

stations_science <- get_station_list(which_api = "science",
station_group = "yellowspot",
state = "nsw",
api_key = mykey)

}
\author{
Rodrigo Pires, \email{rodrigo.pires@dpird.wa.gov.au}
}
