% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_metadata.R
\name{get_station_metadata}
\alias{get_station_metadata}
\title{Get latest BOM station metadata}
\usage{
get_station_metadata()
}
\value{
a \code{data.table} of BOM weather station metadata.
}
\description{
Download the latest station locations and metadata.
}
\details{
If \CRANpkg{ASGS.foyer} is installed locally, this function will
automatically check and correct any invalid state values for stations located
in Australia.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

get_station_metadata()
\dontshow{\}) # examplesIf}
}
\references{
Station location and other metadata are sourced from the Australian Bureau of
Meteorology (\acronym{BOM}) webpage, Bureau of Meteorology Site Numbers:\cr
\url{http://www.bom.gov.au/climate/cdo/about/site-num.shtml}
}
\author{
Adam H. Sparks, \email{adam.sparks@dpird.wa.gov.au}
}
\concept{BOM}
