% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_metadata.R
\name{get_station_metadata}
\alias{get_station_metadata}
\title{Get latest BOM station metadata}
\usage{
get_station_metadata(check_location = FALSE)
}
\arguments{
\item{check_location}{\code{Boolean}. An optional check to use
\CRANpkg{ASGS.foyer} to double check the station's physical locations and
correct any errors in the state where the station is located.
\CRANpkg{ASGS.foyer} must be installed to use this.}
}
\value{
a \code{data.table} of \acronym{BOM} weather stations' metadata for
stations available from \acronym{SILO}.
}
\description{
Download the latest station locations and metadata for stations in the
\acronym{SILO} data set.
}
\details{
If \CRANpkg{ASGS.foyer} is installed locally, this function will
automatically check and correct any invalid state values for stations located
in Australia.
}
\examples{
\dontrun{
get_station_metadata()
}
}
\references{
Station location and other metadata are sourced from the Australian Bureau of
Meteorology (\acronym{BOM}) webpage, Bureau of Meteorology Site Numbers:\cr
\url{http://www.bom.gov.au/climate/cdo/about/site-num.shtml} and
\url{http://www.bom.gov.au/climate/data/lists_by_element/stations.txt}
}
\seealso{
Other SILO: 
\code{\link{get_silo}()}
}
\author{
Adam H. Sparks, \email{adam.sparks@dpird.wa.gov.au}
}
\concept{SILO}
