% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summaries.R
\name{get_summaries}
\alias{get_summaries}
\title{Fetch weather summary from DPIRD weather API for an individual station}
\usage{
get_summaries(
  site,
  first,
  last = Sys.Date(),
  api_key = api_key,
  interval = "daily"
)
}
\arguments{
\item{site}{A string with the station ID code for the station of interest.}

\item{first}{The date on which the weather data summary will be sourced.
\pkg{wrapique} does its best to determine the date given any format but may
fail if given an unconventional date format.}

\item{last}{The last date for which the data will be sourced. For intervals
less than one day, to get one day of data, last should be the same as first,
but must be explicitly coded, as otherwise it will default to the current
date.}

\item{api_key}{\acronym{API} key from \acronym{DPIRD}
\url{https://www.agric.wa.gov.au/web-apis}.  Defaults to \code{NULL}.}

\item{interval}{Time interval to summarise over.  The default is 'daily',
others are '15min', '30min', 'hourly', 'monthly', 'yearly'. For intervals
shorter than 1 day, time period covered will be midnight to midnight, with
the last time interval being before midnight–hour/minute values are for the
end of the time period. Data for shorter intervals ('15min', '30min') should
be available from January of last year.}
}
\value{
a \code{list} with 3 elements: the station code, station name and a nested
\verb{data frame} with the all summary output as per Weather \acronym{API}
documentation.
}
\description{
Fetch weather summary from DPIRD weather API for an individual station
}
\examples{
# You must have an DPIRD API key to proceed
mykey <- 'dpird_api_key'

# set date interval for yearly request
start.date <- "2015-02-01"

# Use default for end data (current system date)
output <- get_summaries(
           site = "AN001",
           first = start.date,
           api_key = mykey,
           interval = "yearly")

# 15 min interval query, define start and end date
start.date <- "2022-05-01"
end.date <- "2022-05-02"

output <- get_summaries(
           site = "BI",
           first = start.date,
           last = end.date,
           api_key = mykey,
           interval = "15min")

}
\seealso{
Other DPIRD: 
\code{\link{get_extreme_weather}()},
\code{\link{tidy_weather_summary}()}
}
\concept{DPIRD}
