% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_list.R
\name{get_weather_list}
\alias{get_weather_list}
\title{Fetch a list of weather station data from DPIRD Science API}
\usage{
get_weather_list(
  sites,
  first,
  last = Sys.Date(),
  data.type,
  api.key,
  verbose = FALSE
)
}
\arguments{
\item{sites}{A character vector or list of station ID codes from the
\acronym{DPIRD} Science \acronym{API} for which data should be fetched.}

\item{first}{The date on which the weather data should start.}

\item{last}{The last date on which the weather data should end.}

\item{data.type}{...}

\item{api.key}{...}

\item{verbose}{... Defaults to \code{FALSE}.}
}
\value{
A \code{list} object of data frames containing the following columns:
\itemize{
\item date,
\item rainfall (rain),
\item maximum temperature (max_temp),
\item minimum temperature (min_temp) and
\item \acronym{DPIRD}'s Science \acronym{API} 2.0 station code.
}
}
\description{
A wrapper function for \code{FoliarDisease::weather_data} to handle fetching data
for several stations specified in a \code{list} object from the \acronym{DPIRD}
Science \acronym{API} in parallel, see the \strong{Stations} section for further
detail. Weather data returned will always start on Jan. 1 of the current year
and carry to the current day or yesterday dependent on data availability in
the database.
}
\examples{
# set up multisession environment
future::plan("multisession")

w <- get_weather_list()

# revert to sequential processing
future::plan("sequential")

}
\author{
Adam Sparks, \email{adam.sparks@@dpird.wa.gov.au}
}
