% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silo_daily_values.R
\docType{data}
\name{silo_daily_values}
\alias{silo_daily_values}
\title{A List of SILO Daily Weather Values}
\format{
A \link{vector} object of 57 items.
}
\source{
\url{https://www.longpaddock.qld.gov.au/silo/about/climate-variables/}
}
\usage{
silo_daily_values
}
\description{
A \link{vector} object containing 18 items representing valid values to supply
to \code{get_patched_point()} and \code{get_data_drill()}'s \var{values} argument
taken from the documentation for the \acronym{SILO} \acronym{API}.
}
\seealso{
Other SILO: 
\code{\link{find_nearby_stations}()},
\code{\link{get_data_drill_apsim}()},
\code{\link{get_data_drill}()},
\code{\link{get_patched_point_apsim}()},
\code{\link{get_patched_point}()},
\code{\link{get_station_metadata}()}

Other data: 
\code{\link{dpird_extreme_weather_values}},
\code{\link{dpird_minute_values}},
\code{\link{dpird_summary_values}}
}
\concept{SILO}
\concept{data}
\keyword{datasets}
