% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_forecast_towns.R
\name{update_forecast_towns}
\alias{update_forecast_towns}
\title{Update internal database with latest BOM forecast towns}
\usage{
update_forecast_towns()
}
\value{
An updated internal package database of \acronym{BOM} précis forecast
towns
}
\description{
Download the latest select forecast towns from the \acronym{BOM} server and
update internal database of précis forecast town names and \acronym{AAC}
codes used by \code{\link{get_precis_forecast}}.  There is no need to use
this unless you know that a forecast town exists in a more current version of
the \acronym{BOM} précis forecast town name database that is not available in
the database distributed with \pkg{wrapique}.  In fact, for
reproducibility purposes, users are discouraged from using this function.
Ported from \pkg{bomrang}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

update_forecast_towns()
\dontshow{\}) # examplesIf}
}
\references{
Data are sourced from: Australian Bureau of Meteorology (\acronym{BOM})
webpage, \dQuote{Weather Data Services},
\url{http://www.bom.gov.au/catalogue/data-feeds.shtml}.
}
\seealso{
Other bomrang-ported: 
\code{\link{find_forecast_towns}()},
\code{\link{get_available_radar}()},
\code{\link{get_coastal_forecast}()},
\code{\link{get_precis_forecast}()},
\code{\link{get_satellite_imagery}()},
\code{\link{parse_precis_forecast}()},
\code{\link{update_station_locations}()}
}
\author{
Adam H. Sparks, \email{adam.sparks@dpird.wa.gov.au}
}
\concept{bomrang-ported}
